% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_stat_dependency.R
\name{add_stat_dependency}
\alias{add_stat_dependency}
\title{add_stat_dependency}
\usage{
add_stat_dependency(calc_dat, confidence_limit = 0.98,
  theoretical = FALSE, relative = TRUE)
}
\description{
Returns relation with confidence limits for each peptide
}
