% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{assert_start_end}
\alias{assert_start_end}
\title{Functions to help with checks}
\usage{
assert_start_end(start, end)
}
\description{
assert_start_end
This function asserts values of parameters. Start (of the sequence) should be 
bigger than 0 and smaller than end of the sequence. The end of the sequence 
should be bigger than start.
}
