% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{assert_time}
\alias{assert_time}
\title{assert_time
This function asserts that time_0 is bigger than 0 and smaller than time_100 
and time_100 is bigger than time_0.}
\usage{
assert_time(time_0, time_100)
}
\description{
assert_time
This function asserts that time_0 is bigger than 0 and smaller than time_100 
and time_100 is bigger than time_0.
}
