% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_aggregated_test_results.R
\name{calculate_aggregated_test_results}
\alias{calculate_aggregated_test_results}
\title{Agregate test result}
\usage{
calculate_aggregated_test_results(
  p_diff_uptake_conf_dat,
  method = c("significance", "weiss"),
  time_t = 1,
  skip_amino = 1
)
}
\arguments{
\item{p_diff_uptake_conf_dat}{...}

\item{time_t}{...}

\item{skip_amino}{\code{integer}, indicator
how many aminos from the N-termins should be 
omitted in visualization}
}
\value{
a \code{\link{data.frame}} object
}
\description{
...
}
\details{
Only peptides without modification are 
aggregated.
}
\examples{
dat <- read_hdx(system.file(package = "HaDeX", "HaDeX/data/KD_180110_CD160_HVEM.csv"))
p_diff_uptake_dat <- create_p_diff_uptake_dataset(dat)
p_diff_uptake_conf_dat <- create_p_diff_uptake_dataset_with_confidence(p_diff_uptake_dat)
calculate_aggregated_test_results(p_diff_uptake_conf_dat, method = "significance")
calculate_aggregated_test_results(p_diff_uptake_conf_dat, method = "weiss")

}
\references{
...
}
\seealso{
\code{\link{read_hdx}}
}
