% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_auc.R
\name{calculate_auc}
\alias{calculate_auc}
\title{Calculate Area Under the Curve for deuterium uptake curves}
\usage{
calculate_auc(
  uptake_dat,
  protein = uptake_dat[["Protein"]][1],
  state = uptake_dat[["State"]][1],
  sequence = uptake_dat[["Sequence"]][1]
)
}
\arguments{
\item{uptake_dat}{}
}
\description{
This function calculates area under the deuterium uptake curve.
}
\details{
The AUC is calculated on the data standarized to unit square by 
division by maximum values of exposure time and deuterium uptake, 
respectively.
}
\examples{
dat <- read_hdx(system.file(package = "HaDeX", "HaDeX/data/KD_180110_CD160_HVEM.csv"))
uptake_dat <- create_uptake_dataset(dat)
calculate_auc(uptake_dat)

}
