% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{calculate_diff_uptake}
\alias{calculate_diff_uptake}
\title{Calculate differential uptake}
\usage{
calculate_diff_uptake(
  dat,
  protein = unique(dat[["Protein"]][1]),
  states = unique(dat[["State"]])[1:2],
  time_0 = 0.001,
  time_t = 1,
  time_100 = 1440,
  deut_part = 0.9
)
}
\arguments{
\item{dat}{data imported by the \code{\link{read_hdx}} function.}

\item{protein}{chosen protein.}

\item{states}{vector of two states for chosen protein. Order is important, as the
deuterium uptake difference is calculated as state_1 - state_2.}

\item{time_0}{minimal exchange control time point of measurement.}

\item{time_t}{time point of the measurement for which the calculations
are done.}

\item{time_100}{maximal exchange control time point of measurement.}

\item{deut_part}{deuterium percentage in solution used in experiment,
value from range [0, 1].}
}
\value{
a \code{\link{data.frame}} object.
}
\description{
Calculate differential uptake
}
\details{
Function \code{\link{calculate_diff_uptake}} calculates
differential values based on provided criteria for peptides for chosen
protein in selected states. The methods of calculation of deuterium uptake
difference, fractional deuterium uptake difference with respect to
minimal/maximal exchange controls or theoretical tabular values are
thoroughly described in the `Data processing` article, as well as
law of propagation of uncertainty, used to calculate uncertainty.
}
\examples{
dat <- read_hdx(system.file(package = "HaDeX", "HaDeX/data/KD_180110_CD160_HVEM.csv"))
diff_dat <- calculate_diff_uptake(dat)
head(diff_dat)

}
\seealso{
\code{\link{read_hdx}}
\code{\link{calculate_state_uptake}}
}
