% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinetics.R
\name{calculate_peptide_kinetics}
\alias{calculate_peptide_kinetics}
\title{Calculate kinetics data for a peptide for a set of states}
\usage{
calculate_peptide_kinetics(
  dat,
  protein,
  sequence,
  states,
  start,
  end,
  time_0,
  time_100,
  deut_part = 0.9
)
}
\arguments{
\item{dat}{dat data imported by the \code{\link{read_hdx}} function.}

\item{protein}{chosen protein.}

\item{sequence}{sequence of chosen peptide.}

\item{states}{vector of states (for chosen protein), for which the 
calculations are done.}

\item{start}{start position of chosen peptide.}

\item{end}{end position of chosen peptide.}

\item{time_0}{minimal exchange control time point of measurement.}

\item{time_100}{maximal exchange control time point of measurement.}

\item{deut_part}{deuterium percentage in solution used in experiment, 
value from range [0, 1].}
}
\description{
...
}
\details{
Function \code{\link{calculate_peptide_kinetics}} calculates
kinetic data for chosen peptide in chosen biological states.
It is a wrapper for \code{\link{calculate_kinetics}} but for mutltiple
states.
The output of this function can be visualized using \code{\link{plot_kinetics}}.
IMPORTANT! The kinetic data is often described as deuterium uptake curve data. 
We use this terms interchangeable.
}
\examples{
dat <- read_hdx(system.file(package = "HaDeX", "HaDeX/data/KD_180110_CD160_HVEM.csv"))

calculate_peptide_kinetics(dat, 
                           protein = "db_CD160",
                           sequence = "INITSSASQEGTRLN", 
                           states = c("CD160", "CD160_HVEM"),
                           start = 1, 
                           end = 15,
                           time_0 = 0.001, 
                           time_100 = 1440)

}
\seealso{
\code{\link{calculate_kinetics}}
\code{\link{calculate_state_uptake}}
\code{\link{plot_kinetics}}
}
