% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_state_deuteration.R
\name{calculate_state_deuteration}
\alias{calculate_state_deuteration}
\title{Calculate deuteration}
\usage{
calculate_state_deuteration(
  dat,
  protein,
  state,
  time_0,
  time_t,
  time_100,
  deut_part = 1
)
}
\arguments{
\item{dat}{data as imported by the \code{\link{read_hdx}} function}

\item{protein}{protein included in calculations}

\item{state}{state included in calculations}

\item{time_0}{experimental `time_in`}

\item{time_t}{chosen time point}

\item{time_100}{experimental `time_out`}

\item{deut_part}{percentage of deuterium the protein was exposed to, value in range [0, 1]}
}
\value{
a \code{\link{data.frame}} object
}
\description{
Calculates deuteration uptake based on supplied parameters.
}
\details{
The function \code{calculate_state_deuteration} calculates deuterium uptake for peptides in given protein in given state based
on supplied parameters: `time_0`, `time_100` and `time_t`. All four variants (combinations of theoretical & relative) are 
supplied (mean values and uncertainty). Manual correction of percentage of deuterium the protein was exposed to during the exchange
in theoretical calculations is provided. 

Methods of calculation and uncertainty are profoundly discussed in the vignette.
}
\examples{
# load example data
dat <- read_hdx(system.file(package = "HaDeX", "HaDeX/data/KD_180110_CD160_HVEM.csv"))

# calculate deuterium uptake values for state "CD160"
calculate_state_deuteration(dat, protein = "db_CD160", state = "CD160",
                            time_0 = 0.001, time_t = 5.000, time_100 = 1440.000)
                            
# calculate deuterium uptake values for state "CD160_HVEM"
calculate_state_deuteration(dat, protein = "db_CD160", state = "CD160_HVEM",
                            time_0 = 0.001, time_t = 5.000, time_100 = 1440.000)
                            
}
\seealso{
\code{\link{read_hdx}} \code{\link{calculate_confidence_limit_values}} \code{\link{add_stat_dependency}}
}
