% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_aggregated_difference_dataset.R
\name{create_aggregated_diff_uptake_dataset}
\alias{create_aggregated_diff_uptake_dataset}
\title{Calculates aggregated uptake difference for peptide pool}
\usage{
create_aggregated_diff_uptake_dataset(diff_uptake_dat)
}
\arguments{
\item{diff_uptake_dat}{...}
}
\description{
Calculates aggregated uptake difference for peptide pool
}
\examples{
diff_uptake_dat <- create_diff_uptake_dataset(alpha_dat)
create_aggregated_diff_uptake_dataset(diff_uptake_dat)

}
