% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_aggregated_test_results.R
\name{create_aggregated_test_results}
\alias{create_aggregated_test_results}
\title{Agregate test result}
\usage{
create_aggregated_test_results(p_diff_uptake_conf_dat, time_t = 1)
}
\arguments{
\item{p_diff_uptake_conf_dat}{...}
}
\value{
a \code{\link{data.frame}} object
}
\description{
...
}
\details{
Only peptides without modification are 
aggregated.
}
\examples{
dat <- read_hdx(system.file(package = "HaDeX", "HaDeX/data/KD_180110_CD160_HVEM.csv"))
p_diff_uptake_dat <- create_p_diff_uptake_dataset(dat)
p_diff_uptake_conf_dat <- create_p_diff_uptake_dataset_with_confidence(p_diff_uptake_dat)
create_aggregated_test_results(p_diff_uptake_conf_dat)

}
\seealso{
\code{\link{read_hdx}}
}
