% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_aggregated_uptake_dataset.R
\name{create_aggregated_uptake_dataset}
\alias{create_aggregated_uptake_dataset}
\title{Calculates the aggregated uptake for peptide pool}
\usage{
create_aggregated_uptake_dataset(kin_dat)
}
\arguments{
\item{kin_dat}{...}
}
\description{
Calculates the aggregated uptake for peptide pool
}
\examples{
kin_dat <- create_uptake_dataset(alpha_dat, states = "Alpha_KSCN")
create_aggregated_uptake_dataset(kin_dat)

}
