% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinetics.R
\name{create_kinetic_dataset}
\alias{create_kinetic_dataset}
\title{Create kinetics dataset for a list of peptides and their states}
\usage{
create_kinetic_dataset(dat, peptide_list, protein, time_0, time_100, deut_part)
}
\arguments{
\item{dat}{dat data imported by the \code{\link{read_hdx}} function.}

\item{peptide_list}{...}

\item{protein}{chosen protein.}

\item{time_0}{minimal exchange control time point of measurement.}

\item{time_100}{maximal exchange control time point of measurement.}

\item{deut_part}{deuterium percentage in solution used in experiment, 
value from range [0, 1].}
}
\value{
a \code{\link{data.frame}} object.
}
\description{
Generates the data set of deuterium uptake between selected 
time points based on supplied peptide list.
}
\details{
This is a wrapper for \code{\link{calculate_kinetics}}, but for
the peptide list instead of one peptide.
}
\examples{
dat <- read_hdx(system.file(package = "HaDeX", "HaDeX/data/KD_180110_CD160_HVEM.csv"))
...

}
\seealso{
\code{\link{calculate_kinetics}}
\code{\link{calculate_state_uptake}}
\code{\link{plot_kinetics}}
}
