% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_p_diff_uptake_dataset_with_confidence.R
\name{create_p_diff_uptake_dataset_with_confidence}
\alias{create_p_diff_uptake_dataset_with_confidence}
\title{Create differential dataset with statistical validity}
\usage{
create_p_diff_uptake_dataset_with_confidence(
  p_diff_uptake_dat,
  houde_interval = NULL,
  houde_interval_times = NULL,
  theoretical = FALSE,
  fractional = FALSE
)
}
\arguments{
\item{p_diff_uptake_dat}{...}

\item{houde_interval}{...}

\item{houde_interval_times}{...}

\item{theoretical}{...}

\item{fractional}{...}
}
\value{
a \code{\link{data.frame}} object.
}
\description{
Create differential dataset with statistical validity
}
\details{
...
}
\examples{
dat <- read_hdx(system.file(package = "HaDeX", "HaDeX/data/KD_180110_CD160_HVEM.csv"))
p_diff_uptake_dat <- create_p_diff_uptake_dataset(dat)
p_diff_uptake_dat_confidence <- create_p_diff_uptake_dataset_with_confidence(p_diff_uptake_dat)
head(p_diff_uptake_dat_confidence)

}
\seealso{
\code{\link{read_hdx}}
\code{\link{create_p_diff_uptake_dataset}}
\code{\link{calculate_confidence_limit_values}}
}
