% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{create_state_comparison_dataset}
\alias{create_state_comparison_dataset}
\title{Creates comparison uptake dataset}
\usage{
create_state_comparison_dataset(
  dat,
  protein = unique(dat[["Protein"]])[1],
  states = unique(dat[["State"]]),
  time_0 = 0.001,
  time_t = 1,
  time_100 = 1440,
  deut_part = 0.9
)
}
\arguments{
\item{dat}{data imported by the \code{\link{read_hdx}} function.}

\item{protein}{chosen protein.}

\item{states}{vector of states (for chosen protein), for which the
calculations are done.}

\item{time_0}{minimal exchange control time point of measurement.}

\item{time_t}{time point of the measurement for which the calculations
are done.}

\item{time_100}{maximal exchange control time point of measurement.}

\item{deut_part}{deuterium percentage in solution used in experiment,
value from range [0, 1].}
}
\value{
a \code{\link{data.frame}} object.
}
\description{
Creates comparison uptake dataset
}
\details{
Function \code{\link{create_state_comparison_dataset}} is a
wrapper for \code{\link{calculate_state_uptake}} function, calls
this function for all (default) or chosen states in states vector.
}
\examples{
dat <- read_hdx(system.file(package = "HaDeX", "HaDeX/data/KD_180110_CD160_HVEM.csv"))
comparison_dat <- create_state_comparison_dataset(dat)
head(comparison_dat)

}
\seealso{
\code{\link{read_hdx}}
\code{\link{calculate_state_uptake}}
}
