% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_state_uptake_dataset.R
\name{create_state_uptake_dataset}
\alias{create_state_uptake_dataset}
\title{Create uptake dataset for chosen state}
\usage{
create_state_uptake_dataset(
  dat,
  protein = unique(dat[["Protein"]])[1],
  state = (dat[["State"]])[1],
  time_0 = min(dat[dat[["Exposure"]] > 0, ][["Exposure"]]),
  time_100 = max(dat[["Exposure"]]),
  deut_part = 0.9
)
}
\arguments{
\item{dat}{data imported by the \code{\link{read_hdx}} function.}

\item{protein}{chosen protein.}

\item{state}{biological state for chosen protein.}

\item{time_0}{minimal exchange control time point of measurement [min].}

\item{time_100}{maximal exchange control time point of measurement [min].}

\item{deut_part}{deuterium percentage in solution used in experiment, 
value from range [0, 1].}
}
\value{
a \code{\link{data.frame}} object.
}
\description{
Calculates deuterium uptake values for one 
biological state.
}
\details{
The function \code{\link{create_uptake_dataset}} generates 
a dataset with deuterium uptake values in different forms. For each
peptide in chosen protein in chosen state for time points of measurement
between minimal and maximal control time points of measurement deuterium 
uptake, fractional deuterium uptake with respect to controls or theoretical
tabular values are calculated, with combined and propagated uncertainty. 
Each peptide has an ID, based on its start position. 
This data can be presented in a form of comparison plot, butterfly plot or
chiclet plot.
}
\examples{
dat <- read_hdx(system.file(package = "HaDeX", "HaDeX/data/KD_180110_CD160_HVEM.csv"))
state_uptake_dat <- create_state_uptake_dataset(dat)
head(state_uptake_dat)

}
\seealso{
\code{\link{read_hdx}}
\code{\link{calculate_state_uptake}}
}
