% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_butterfly_elements.R
\name{generate_butterfly_data}
\alias{generate_butterfly_data}
\title{generate_butterfly_data}
\usage{
generate_butterfly_data(butterfly_dat, theoretical = FALSE, fractional = FALSE)
}
\arguments{
\item{butterfly_dat}{data as imported by the \code{\link{read_hdx}} function}

\item{theoretical}{\code{logical}, determines if values are theoretical}

\item{fractional}{\code{logical}, determines if values are fractional}
}
\value{
...
}
\description{
Generates butterfly data, based on the supplied
parameters.
}
\details{
This data is available in the GUI. 
All of the numerical values are rounded to 4 places after the dot!!
}
\seealso{
...
}
