% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_butterfly_elements.R
\name{generate_butterfly_dataset}
\alias{generate_butterfly_dataset}
\title{generate_butterfly_dataset}
\usage{
generate_butterfly_dataset(
  dat,
  protein = unique(dat[["Protein"]])[1],
  state = (dat[["State"]])[1],
  time_0 = 0.001,
  time_100 = 1440,
  deut_part = 1
)
}
\arguments{
\item{dat}{data as imported by the \code{\link{read_hdx}} function}

\item{protein}{...}

\item{state}{...}

\item{time_0}{...}

\item{time_100}{...}

\item{deut_part}{...}
}
\value{
...
}
\description{
generate_butterfly_dataset
}
\details{
...
}
\seealso{
...
}
