% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_butterfly_differential_elements.R
\name{generate_butterfly_differential_data}
\alias{generate_butterfly_differential_data}
\title{Generate butterfly differential data}
\usage{
generate_butterfly_differential_data(
  diff_uptake_dat,
  theoretical = FALSE,
  fractional = FALSE
)
}
\arguments{
\item{diff_uptake_dat}{data produced by 
\code{\link{create_diff_uptake_dataset}} function.}

\item{theoretical}{\code{logical}, determines if values are theoretical.}

\item{fractional}{\code{logical}, determines if values are fractional.}
}
\value{
a \code{\link{data.frame}} object.
}
\description{
Generate butterfly differential data
}
\details{
This function subsets the dataset based on provided criteria,
rounds the numerical values (4 places) and changes the column names 
to user-friendly ones. 
This data is available in the GUI.
}
\examples{
dat <- read_hdx(system.file(package = "HaDeX", "HaDeX/data/KD_180110_CD160_HVEM.csv"))
diff_uptake_dat <- create_diff_uptake_dataset(dat)
head(generate_butterfly_differential_data(diff_uptake_dat))

}
\seealso{
\code{\link{read_hdx}}
\code{\link{create_diff_uptake_dataset}}
\code{\link{plot_butterfly_differential}}
}
