% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_butterfly_differential_elements.R
\name{generate_butterfly_differential_dataset}
\alias{generate_butterfly_differential_dataset}
\title{generate_butterfly_differential_dataset}
\usage{
generate_butterfly_differential_dataset(
  dat,
  protein = unique(dat[["Protein"]])[1],
  state_1 = unique(dat[["State"]])[1],
  state_2 = unique(dat[["State"]])[2],
  time_0 = 0.001,
  time_100 = 1440,
  deut_part = 1
)
}
\arguments{
\item{dat}{...}

\item{protein}{...}

\item{state_1}{...}

\item{state_2}{...}

\item{time_0}{...}

\item{time_100}{...}

\item{deut_part}{...}
}
\value{

}
\description{
generate_butterfly_differential_dataset
}
\details{

}
\seealso{

}
