% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_chiclet_elements.R
\name{generate_chiclet_plot}
\alias{generate_chiclet_plot}
\title{generate_chiclet_plot}
\usage{
generate_chiclet_plot(
  chiclet_dat,
  theoretical = FALSE,
  fractional = FALSE,
  show_uncertainty = FALSE
)
}
\arguments{
\item{chiclet_dat}{produced by generate_butterfly_dataset}

\item{theoretical}{\code{logical}, determines if values are theoretical}

\item{fractional}{\code{logical}, determines if values are fractional}

\item{show_uncertainty}{...}
}
\value{
...
}
\description{
Generates chiclet plot based on supplied data
and parameters.
}
\details{
This plot is visible in GUI.
}
\seealso{
...
}
