% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_comparison_plot.R
\name{generate_comparison_plot}
\alias{generate_comparison_plot}
\title{generate_comparison_plot}
\usage{
generate_comparison_plot(dat, theoretical, relative)
}
\arguments{
\item{dat}{produced by \code{\link{calculate_state_deuteration}} function}

\item{theoretical}{...}

\item{relative}{...}
}
\value{
...
}
\description{
Generates comparison plot based on supplied data
and parameters.
}
\details{
This plot is visible in GUI. The names of the parameters
and variables will be changed later after the glossary project.
}
\seealso{
...
}
