% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_differential_data.R
\name{generate_differential_data}
\alias{generate_differential_data}
\title{generate_differential_data}
\usage{
generate_differential_data(
  dat,
  states,
  protein,
  time_in,
  time_chosen,
  time_out,
  deut_part
)
}
\arguments{
\item{dat}{...}

\item{states}{vector of two states to calculate difference between them, 
the order is important}

\item{protein}{...}

\item{time_in}{...}

\item{time_chosen}{...}

\item{time_out}{...}

\item{deut_part}{...}
}
\value{
...
}
\description{
Generate the data frame with differential data from two
provided states - experimental/relative calculations with the
uncertainty
}
\details{
The names of the parameters will be changed to more descriptive 
ones after the glossary project.
}
\seealso{
...
}
