% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_differential_data_set.R
\name{generate_differential_data_set}
\alias{generate_differential_data_set}
\title{generate_differential_data_set}
\usage{
generate_differential_data_set(
  dat,
  states,
  protein,
  time_in,
  time_chosen,
  time_out,
  deut_part
)
}
\arguments{
\item{dat}{...}

\item{states}{vector of two states to calculate difference between them, 
the order is important}

\item{protein}{...}

\item{time_in}{...}

\item{time_chosen}{...}

\item{time_out}{...}

\item{deut_part}{...}
}
\value{
...
}
\description{
Generate the data frame with differential data from two
provided states - experimental/relative calculations with the
uncertainty, based on supplied parameters.
}
\details{
The names of the parameters and variables will be changed 
later after the glossary project.
}
\seealso{
...
}
