% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_general_data_set.R
\name{generate_general_data_set}
\alias{generate_general_data_set}
\title{generate_general_data_set}
\usage{
generate_general_data_set(
  dat,
  control_protein,
  control_state,
  control_exposure
)
}
\arguments{
\item{dat}{...}

\item{control_protein}{...}

\item{control_state}{...}

\item{control_exposure}{...}
}
\value{
...
}
\description{
Generates the data set with control sample, based on 
supplied parameters.
}
\details{
The names of the parameters and variables will be changed 
later after the glossary project.
}
\seealso{
...
}
