% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_general_dataset.R
\name{generate_general_dataset}
\alias{generate_general_dataset}
\title{generate_general_dataset}
\usage{
generate_general_dataset(dat, control_protein, control_state, control_exposure)
}
\arguments{
\item{dat}{...}

\item{control_protein}{...}

\item{control_state}{...}

\item{control_exposure}{...}
}
\value{
...
}
\description{
Generates the dataset with control sample 
based on supplied parameters.
}
\details{
...
}
\seealso{
...
}
