% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_kinetic_data_set.R
\name{generate_kinetic_data_set}
\alias{generate_kinetic_data_set}
\title{generate_kinetic_data_set}
\usage{
generate_kinetic_data_set(
  dat,
  peptide_list,
  protein,
  time_in,
  time_out,
  deut_concentration
)
}
\arguments{
\item{dat}{...}
}
\value{
...
}
\description{
...
}
\details{
...
}
\seealso{
...
}
