% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_kinetic_plot.R
\name{generate_kinetic_plot}
\alias{generate_kinetic_plot}
\title{generate_kinetic_plot}
\usage{
generate_kinetic_plot(dat, theoretical, relative)
}
\arguments{
\item{dat}{produced by \code{\link{generate_kinetic_data_set}} function}

\item{theoretical}{...}

\item{relative}{...}
}
\value{
...
}
\description{
Generates deuterium uptake curve based on supplied data
and parameters.
}
\details{
This plot is visible in GUI.
}
\seealso{
...
}
