% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_overlap_elements.R
\name{generate_overlap_data}
\alias{generate_overlap_data}
\title{generate_overlap_data}
\usage{
generate_overlap_data(dat, protein, state, start, end)
}
\arguments{
\item{dat}{...}

\item{protein}{...}

\item{state}{...}

\item{start}{...}

\item{end}{...}
}
\value{
...
}
\description{
Generates overlap data, based on the supplied
parameters.
}
\details{
This data is available in the GUI.
}
\seealso{
...
}
