% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_overlap_distribution_plot.R
\name{generate_overlap_distribution_plot}
\alias{generate_overlap_distribution_plot}
\title{generate_overlap_distribution_plot}
\usage{
generate_overlap_distribution_plot(dat, start, end)
}
\arguments{
\item{dat}{data frame generated by \code{\link{generate_overlap_distribution_plot}}
function}

\item{start}{...}

\item{end}{...}
}
\value{
...
}
\description{
Generates the frequency of overlap plot.
}
\details{
This plot is visible in the GUI.
}
\seealso{
...
}
