% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_quality_control_data.R
\name{generate_quality_control_data}
\alias{generate_quality_control_data}
\title{generate_quality_control_data}
\usage{
generate_quality_control_data(dat)
}
\arguments{
\item{dat}{data frame from \code{\link{quality_control}}, 
scaled if necessary.}
}
\value{
...
}
\description{
Generates the quality control data in user-friendly way.
}
\details{
This data is available from the GUI. May be internal.
}
\seealso{
...
}
