% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_quality_control_plot.R
\name{generate_quality_control_plot}
\alias{generate_quality_control_plot}
\title{generate_quality_control_plot}
\usage{
generate_quality_control_plot(dat)
}
\arguments{
\item{dat}{data frame from \code{\link{quality_control}}, 
scaled if necessary.}
}
\value{
...
}
\description{
Generates quality control plot.
}
\details{
This plot is visible in the GUI.
}
\seealso{
...
}
