% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_summary_table.R
\name{generate_summary_table}
\alias{generate_summary_table}
\title{generate_summary_table}
\usage{
generate_summary_table(
  dat,
  confidence_limit_1,
  confidence_limit_2,
  overlap_distribution_data
)
}
\arguments{
\item{dat}{...}

\item{confidence_limit_1}{...}

\item{confidence_limit_2}{...}

\item{overlap_distribution_data}{custom format, generated by
\code{\link{generate_overlap_distribution_data}}}
}
\value{
...
}
\description{
Generates summary table.
}
\details{
The format in the table is suggested in Masson, G.R., Burke, J.E., 
Ahn, N.G., Anand, G.S., Borchers, C., Brier, S., Bou-Assaf, G.M., Engen, J.R., 
Englander, S.W., Faber, J., et al. (2019). Recommendations for performing, 
interpreting and reporting hydrogen deuterium exchange mass spectrometry 
(HDX-MS) experiments. Nat Methods 16, 595–602
}
\seealso{
...
}
