% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_volcano_elements.R
\name{generate_volcano_data}
\alias{generate_volcano_data}
\title{generate_volcano_data}
\usage{
generate_volcano_data(vol_data)
}
\arguments{
\item{vol_data}{data produced by the \code{\link{generate_volcano_dataset}} 
function.}
}
\description{
Generates a nice data frame
}
\details{
This function rounds the numerical values (4 places) and 
changes the column names to user friendly ones. 
This data is available in the GUI.
}
\examples{
dat <-  dat <- read_hdx(system.file(package = "HaDeX", "HaDeX/data/KD_180110_CD160_HVEM.csv"))
vol_dat <- generate_volcano_dataset(dat)
generate_volcano_data(vol_dat, show_confidence_limits = T)

}
\seealso{
\code{\link{generate_volcano_dataset}} \code{\link{generate_volcano_plot}}
}
