% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_volcano_elements.R
\name{generate_volcano_dataset}
\alias{generate_volcano_dataset}
\title{generate_volcano_data}
\usage{
generate_volcano_dataset(
  dat,
  protein = unique(dat[["Protein"]])[1],
  state_1 = unique(dat[["State"]])[1],
  state_2 = unique(dat[["State"]])[2]
)
}
\arguments{
\item{dat}{...}

\item{protein}{...}

\item{states}{...}
}
\description{
Generates data set for volcano plot
}
