% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_vulcano_elements.R
\name{generate_vulcano_data}
\alias{generate_vulcano_data}
\title{generate_vulcano_data}
\usage{
generate_vulcano_data(
  dat,
  protein = unique(dat[["Protein"]])[1],
  states = unique(dat[["State"]])[1:2]
)
}
\arguments{
\item{dat}{...}

\item{protein}{...}

\item{states}{...}
}
\description{
Generates data for vulcano plot
}
