% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_vulcano_elements.R
\name{generate_vulcano_plot}
\alias{generate_vulcano_plot}
\title{generate_vulcano_plot}
\usage{
generate_vulcano_plot(vul_data, state_1 = "", state_2 = "")
}
\arguments{
\item{vul_data}{...}
}
\description{
Generates vulcano plot based on suplied vulcano data
}
