% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_woods_data.R
\name{generate_woods_data}
\alias{generate_woods_data}
\title{generate_woods_data}
\usage{
generate_woods_data(
  dat,
  theoretical,
  relative,
  confidence_limit_1,
  confidence_limit_2
)
}
\arguments{
\item{dat}{custom format, produced by \code{\link{generate_differential_data}}}

\item{theoretical}{...}

\item{relative}{...}

\item{confidence_limit_1}{...}

\item{confidence_limit_2}{...}
}
\value{
...
}
\description{
Generates Woods Plot data, compatible 
with the plot.
}
\details{
This data is accessible from the GUI.
}
\seealso{
...
}
