% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_woods_plot.R
\name{generate_woods_plot}
\alias{generate_woods_plot}
\title{generate_woods_plot}
\usage{
generate_woods_plot(
  dat,
  theoretical,
  relative,
  confidence_limit,
  confidence_limit_2
)
}
\arguments{
\item{dat}{custom format}

\item{theoretical}{...}

\item{relative}{...}

\item{confidence_limit}{...}

\item{confidence_limit_2}{...}
}
\value{
...
}
\description{
Generates Woods Plot with confidence values based 
on supplied parameters
}
\details{
This plot is visible in GUI.
}
\seealso{
...
}
