% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{get_n_replicates}
\alias{get_n_replicates}
\title{Get number of replicates}
\usage{
get_n_replicates(dat, protein = dat[["Protein"]][1])
}
\arguments{
\item{dat}{data imported by the \code{\link{read_hdx}} function.}
}
\value{
a \code{\link{numeric}} value.
}
\description{
Get number of replicates
}
\examples{
dat <- read_hdx(system.file(package = "HaDeX", "HaDeX/data/KD_180110_CD160_HVEM.csv"))
get_n_replicates(dat)

}
\seealso{
\code{\link{read_hdx}}
}
