% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{get_peptide_sequence}
\alias{get_peptide_sequence}
\title{Get peptide sequence based on the position}
\usage{
get_peptide_sequence(
  dat,
  protein = dat[["Protein"]][1],
  start,
  end,
  modification = FALSE
)
}
\arguments{
\item{dat}{any data frame that contains following information:
protein, sequence, start, end, modification.}

\item{protein}{...}

\item{start}{start position of the peptide of interest.}

\item{end}{end position of the peptide of interest.}

\item{modification}{logical value to indicate if peptide
of interest has modification or not.}
}
\value{
character value.
}
\description{
Get peptide sequence based on the position
}
\examples{
dat <- read_hdx(system.file(package = "HaDeX", "HaDeX/data/KD_180110_CD160_HVEM.csv"))
get_peptide_sequence(dat, start = 1, end = 15)

}
\seealso{
\code{\link{read_hdx}}
}
