% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{get_protein_redundancy}
\alias{get_protein_redundancy}
\title{Get protein redundancy}
\usage{
get_protein_redundancy(dat, protein_length = NULL)
}
\arguments{
\item{dat}{data imported by the \code{\link{read_hdx}} function.}

\item{protein_length}{\code{\link{numeric}}, indicates the length of 
the protein. If not provided, the maximal end value from the file is used.}
}
\value{
a \code{\link{numeric}} value.
}
\description{
Get protein redundancy
}
\details{
Function \code{\link{get_protein_redundancy}} calculates the redundancy
of the protein, based on provided experimental data.
}
\examples{
dat <- read_hdx(system.file(package = "HaDeX", "HaDeX/data/KD_180110_CD160_HVEM.csv"))
get_protein_redundancy(dat)

}
\seealso{
\code{\link{read_hdx}}
}
