% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_reconstruced_sequence.R
\name{get_reconstructed_sequence}
\alias{get_reconstructed_sequence}
\title{get_reconstruced_sequence}
\usage{
get_reconstructed_sequence(dat)
}
\description{
Reconstructs sequence from DynamiX file. If a position is not covered by experiment, x is shown.
}
