% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphic_overlapping.R
\name{graphic_overlapping}
\alias{graphic_overlapping}
\title{Plot graphic sequence overlapping}
\usage{
graphic_overlapping(dat, chosen_state)
}
\arguments{
\item{dat}{data as imported by the \code{\link{read_hdx}} function}

\item{chosen_state}{sequence states to be included in plot}
}
\description{
Produces graphic sequence overlapping plot based on experimental HDX-MS data.
}
\examples{
dat <- read_hdx(system.file(package = "HaDeX", 
                            "HaDeX/data/KD_180110_CD160_HVEM.csv"))
graphic_overlapping(dat, "CD160_HVEM")

}
