% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HaDeXify.R
\name{HaDeXify}
\alias{HaDeXify}
\title{HaDeX customized ggplot theme}
\usage{
HaDeXify(plt)
}
\arguments{
\item{plt}{ggplot object. Plot to HaDeXify.}
}
\description{
This function HaDeXifies plot. It adds HaDeX logo and ggplot theme.
}
\examples{
dat <- read_hdx(system.file(package = "HaDeX", "HaDeX/data/KD_180110_CD160_HVEM.csv"))
diff_uptake_dat <- generate_differential_data_set(dat)
HaDeXify(plot_differential(diff_uptake_dat))
}
