% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdx_data_class.R
\name{hdx_data}
\alias{hdx_data}
\title{Creation of validated hdx_data class}
\usage{
hdx_data(dat, source, has_modification, msg = "")
}
\arguments{
\item{dat}{\code{\link{data.frame}}, read from the datafile.}

\item{source}{\code{\link{character}}, the source of the datafile.}

\item{has_modification}{\code{\link{logical}}, indicator if there are 
modified peptides in the datafile.}
}
\value{
\code{\link{hdx_data}} object.
}
\description{
The wrapper function for the constructor of the \code{\link{hdx_data}} and 
its validator. Used in \code{\link{read_hdx}} function.
}
\keyword{internal}
