% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HaDeX_gui.R
\name{is_GUI_installed}
\alias{is_GUI_installed}
\title{Checks if GUI package is installed}
\usage{
is_GUI_installed()
}
\value{
logical value indicating availability of GUI package
}
\description{
Indicates presence or absence of HaDeXGUI package.
}
