% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_aggregated_difference.R
\name{plot_aggregated_differential_uptake}
\alias{plot_aggregated_differential_uptake}
\title{Plots aggregated uptake difference}
\usage{
plot_aggregated_differential_uptake(
  aggregated_diff_dat,
  fractional = TRUE,
  theoretical = FALSE,
  time_100 = max(unique(aggregated_diff_dat[["Exposure"]])),
  panels = FALSE,
  interactive = FALSE
)
}
\arguments{
\item{aggregated_diff_dat}{...}

\item{fractional}{...}

\item{theoretical}{...}

\item{time_100}{...}

\item{panels}{...}

\item{interactive}{...}
}
\description{
Plots aggregated uptake difference
}
\examples{
diff_uptake_dat <- create_diff_uptake_dataset(alpha_dat)
aggregated_diff_dat <- create_aggregated_diff_uptake_dataset(diff_uptake_dat)
plot_aggregated_differential_uptake(aggregated_diff_dat, panels = F)
plot_aggregated_differential_uptake(aggregated_diff_dat, fractional = F, theoretical = T, panels = F)
plot_aggregated_differential_uptake(aggregated_diff_dat, theoretical = T, panels = F)

}
