% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_aggregated_uptake.R
\name{plot_aggregated_uptake}
\alias{plot_aggregated_uptake}
\title{Plots aggregated uptake}
\usage{
plot_aggregated_uptake(
  aggregated_dat,
  fractional = TRUE,
  theoretical = FALSE,
  time_100 = max(unique(aggregated_dat[["Exposure"]])),
  panels = FALSE,
  interactive = FALSE
)
}
\arguments{
\item{aggregated_dat}{...}

\item{fractional}{...}

\item{theoretical}{...}

\item{time_100}{...}

\item{panels}{...}

\item{interactive}{...}
}
\description{
Plots aggregated uptake
}
\examples{
kin_dat <- create_uptake_dataset(alpha_dat, states = "Alpha_KSCN")
aggregated_dat <- create_aggregated_uptake_dataset(kin_dat)
plot_aggregated_uptake(aggregated_dat, panels = F)
plot_aggregated_uptake(aggregated_dat, fractional = F, panels = F)
plot_aggregated_uptake(aggregated_dat, fractional = F, theoretical = T, panels = T)

}
