% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_aggregated_uptake_structure.R
\name{plot_aggregated_uptake_structure}
\alias{plot_aggregated_uptake_structure}
\title{Plot aggregated uptake on the 3D structure}
\usage{
plot_aggregated_uptake_structure(
  aggregated_dat,
  differential = FALSE,
  fractional = TRUE,
  theoretical = FALSE,
  time_t,
  pdb_file_path
)
}
\description{
Plot aggregated uptake on the 3D structure
}
\examples{
diff_uptake_dat <- create_diff_uptake_dataset(alpha_dat)
diff_aggregated_dat <- create_aggregated_diff_uptake_dataset(diff_uptake_dat)

kin_dat <- create_uptake_dataset(alpha_dat, states = "Alpha_KSCN" )
aggregated_dat <- create_aggregated_uptake_dataset(kin_dat)

}
