% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_amino_distribution.R
\name{plot_amino_distribution}
\alias{plot_amino_distribution}
\title{generate_amino_distribution}
\usage{
plot_amino_distribution(
  position_in_sequence,
  hydro_properties,
  protein,
  charge_colors,
  interactive = getOption("hadex_use_interactive_plots")
)
}
\arguments{
\item{position_in_sequence}{custom format}

\item{hydro_properties}{...}

\item{protein}{...}

\item{charge_colors}{...}

\item{interactive}{\code{logical}, whether plot should have an interactive 
layer created with with ggiraph, which would add tooltips to the plot in an
interactive display (HTML/Markdown documents or shiny app).}
}
\value{
a \code{\link{ggplot2}} object.
}
\description{
Generates amino distribution based on the 
protein sequence and shows if the amino acid is hydrophobic
or hydrophylic.
}
\details{
The data for this function is not packaged yet.
}
\examples{
...

}
\seealso{
...
}
