% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinetics.R
\name{plot_kinetics}
\alias{plot_kinetics}
\title{Plot kinetics data}
\usage{
plot_kinetics(
  kin_dat,
  theoretical = FALSE,
  fractional = FALSE,
  uncertainty_type = "ribbon",
  log_x = TRUE
)
}
\arguments{
\item{kin_dat}{calculated kinetic data by \code{\link{calculate_kinetics}} 
or \code{\link{calculate_peptide_kinetics}} or \code{\link{create_kinetics_dataset}}
function.}

\item{theoretical}{\code{logical}, determines if plot shows theoretical values.}

\item{fractional}{\code{logical}, determines if plot shows fractional values.}

\item{uncertainty_type}{type of presenting uncertainty, possible values:
"ribbon", "bars" or "bars + line".}

\item{log_x}{\code{logical}, determines if x axis shows logarithmic values.}
}
\value{
a \code{\link{ggplot2}} object.
}
\description{
Plots kinetics of the hydrogen-deuterium exchange for specific peptides.
}
\details{
This function visualizes the output of the  
\code{\link{calculate_kinetics}} function. 
Based on supplied parameters appropriate columns are chosen for the plot. 
The uncertainty associated with each peptide is shown as a ribbon. 
Axis are labeled according to the supplied parameters but no title is provided.

If you want to plot data for more then one peptide in one state, join 
calculated data by using \code{\link{bind_rows}} from dplyr package and 
pass the result as kin_dat.
}
\examples{
dat <- read_hdx(system.file(package = "HaDeX", "HaDeX/data/KD_180110_CD160_HVEM.csv"))

# one peptide in one state
kin1 <- calculate_kinetics(dat, 
                           protein = "db_CD160",
                           sequence = "INITSSASQEGTRLN", 
                           state = "CD160",
                           start = 1, 
                           end = 15,
                           time_in = 0.001, 
                           time_out = 1440)
plot_kinetics(kin_dat = kin1, 
              theoretical = FALSE, 
              fractional = TRUE)

# one peptide in all states         
kin2 <- calculate_peptide_kinetics(dat, 
                                   protein = "db_CD160",
                                   sequence = "INITSSASQEGTRLN", 
                                   states = c("CD160", "CD160_HVEM"),
                                   start = 1, 
                                   end = 15,
                                   time_0 = 0.001, 
                                   time_100 = 1440)
plot_kinetics(kin_dat = kin2, 
              theoretical = FALSE, 
              fractional = TRUE)
                
}
\seealso{
\code{\link{read_hdx}}
\code{\link{calculate_kinetics}}
\code{\link{calculate_peptide_kinetics}}
}
