% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_plots.R
\name{plot_manhattan}
\alias{plot_manhattan}
\title{Manhattan plot}
\usage{
plot_manhattan(
  p_dat,
  plot_title = NULL,
  separate_times = T,
  show_confidence_limit = T
)
}
\arguments{
\item{p_dat}{...}

\item{plot_title}{...}

\item{separate_times}{...}

\item{show_confidence_limit}{...}
}
\value{
...
}
\description{
Manhattan plot
}
\details{
...
}
\examples{
dat <-  read_hdx(system.file(package = "HaDeX", "HaDeX/data/KD_180110_CD160_HVEM.csv"))
p_dat <- create_p_diff_uptake_dataset(dat)
manhattan_plot(p_dat)

manhattan_plot(p_dat, separate_times = F)

manhattan_plot(p_dat, separate_times = F, show_confidence_limit = F)
 
}
\references{
...
}
\seealso{
...
}
