% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlaps.R
\name{plot_overlap}
\alias{plot_overlap}
\title{generate_overlap_plot}
\usage{
plot_overlap(dat)
}
\arguments{
\item{dat}{produced by \code{\link{generate_overlap_plot}}
function}
}
\value{
...
}
\description{
Generates overapping peptide plot based on supplied data
and parameters.
}
\details{
This plot is visible in GUI.
}
\seealso{
...
}
