% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_control.R
\name{plot_quality_control}
\alias{plot_quality_control}
\title{generate_quality_control_plot}
\usage{
plot_quality_control(dat)
}
\arguments{
\item{dat}{data produced by \code{\link{create_quality_control_dataset}} 
function, scaled if necessary.}
}
\value{
a \code{\link{ggplot2}} object.
}
\description{
Generates quality control plot based on supplied data.
}
\details{
This plot is visible in GUI.
}
\seealso{
\code{\link{create_quality_control_dataset}}
\code{\link{show_quality_control_data}}
}
