% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicates.R
\name{plot_replicate_histogram}
\alias{plot_replicate_histogram}
\title{Plot replicates histogram}
\usage{
plot_replicate_histogram(rep_dat)
}
\arguments{
\item{rep_dat}{replicate data, created by 
\code{\link{create_replicate_dataset}} function.}
}
\value{
a \code{\link{ggplot}} object.
}
\description{
Plot histogram on number of replicates per
peptide in one or multiple time point of measurement.
}
\details{
The function shows two versions of replicate 
histogram, based on supplied \code{rep_dat}. 
On the X-axis there are peptide ID, and on the Y-axis
there are numbers of replicates. 
If \code{rep_dat} contains data from one time point 
of measurement, the histogram colors reflect the 
number of replicates to highlight the outliers.
If \code{rep_dat} contains multiple time point of
measurement, the colors help to distinguish between 
them.
}
\examples{
dat <- read_hdx(system.file(package = "HaDeX", "HaDeX/data/KD_180110_CD160_HVEM.csv"))
rep_dat <- create_replicate_dataset(dat)
plot_replicate_histogram(rep_dat)

rep_dat <- create_replicate_dataset(dat, time_t = 0.167)
plot_replicate_histogram(rep_dat)

}
\seealso{
\code{\link{create_replicate_dataset}} 
\code{\link{show_replicate_histogram_data}}
}
