% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_replicate_mass_uptake.R
\name{plot_replicate_mass_uptake}
\alias{plot_replicate_mass_uptake}
\title{Plots the mass uptake curve from the replicates}
\usage{
plot_replicate_mass_uptake(
  dat,
  protein = dat[["Protein"]][1],
  state = dat[["State"]][1],
  sequence = dat[["Sequence"]][1],
  show_aggregated = FALSE
)
}
\arguments{
\item{dat}{...}

\item{protein}{...}

\item{state}{...}

\item{sequence}{...}

\item{show_aggregated}{...}
}
\description{
Plots the mass uptake curve from the replicates
}
\examples{
dat <- read_hdx(system.file(package = "HaDeX", "HaDeX/data/KD_180110_CD160_HVEM.csv"))
plot_replicate_mass_uptake(dat)

}
\seealso{
\code{\link{read_hdx}}
}
