% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uptake_plots.R
\name{plot_state_comparison}
\alias{plot_state_comparison}
\title{Plot state comparison}
\usage{
plot_state_comparison(dat, theoretical = FALSE, fractional = FALSE)
}
\arguments{
\item{dat}{data produced by \code{\link{calculate_state_deuteration}} or
\code{\link{generate_comparison_dataset}} function.}

\item{theoretical}{\code{logical}, determines if values are theoretical.}

\item{fractional}{\code{logical}, determines if values are fractional.}
}
\value{
a \code{ggplot} object.
}
\description{
Plot state comparison
}
\details{
Function \code{\link{plot_comparison}} presents provided 
data in a form of comparison plot, for peptides for chosen protein in chosen states,
at one time point of measurement at once. On X-axis there is a position in a sequence, 
with length of a segment of each peptide representing its length. On Y-axis there 
is deuterium uptake in chosen form. Error bars represents the combined and propagated
uncertainty. 
This plot is visible in GUI.
}
\examples{
dat <- read_hdx(system.file(package = "HaDeX", "HaDeX/data/KD_180110_CD160_HVEM.csv"))
comparison_dat <- generate_comparison_dataset(dat)
plot_state_comparison(comparison_dat)

}
\seealso{
\code{\link{calculate_state_deuteration}}  
\code{\link{generate_comparison_dataset}}
\code{\link{generate_comparison_data}}
}
