% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_uncertainty.R
\name{plot_uncertainty}
\alias{plot_uncertainty}
\title{Plot the uncertainty of the measurement for peptides}
\usage{
plot_uncertainty(
  dat,
  protein = dat[["Protein"]][1],
  state = dat[["State"]][1],
  aggregated = TRUE,
  separate_times = TRUE,
  show_threshold = TRUE
)
}
\arguments{
\item{dat}{...}

\item{protein}{...}

\item{state}{...}

\item{aggregated}{...}

\item{separate_times}{....}

\item{show_threshold}{...}
}
\value{
a \code{\link{ggplot}} object.
}
\description{
Plot the uncertainty of the mass measurements - for aggregated
data or before aggregation - to see if there is a region with uncertainty 
higher than acceptable.
}
\details{
...
}
\examples{
dat <- read_hdx(system.file(package = "HaDeX", "HaDeX/data/KD_180110_CD160_HVEM.csv"))
plot_uncertainty(dat)


}
\seealso{
\code{\link{read_hdx}}
\code{\link{calculate_exp_masses}}
}
