% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_volcano.R
\name{plot_volcano}
\alias{plot_volcano}
\title{Volcano plot}
\usage{
plot_volcano(
  p_dat,
  state_1 = "",
  state_2 = "",
  adjust_axes = TRUE,
  show_confidence_limits = FALSE,
  confidence_level = 0.98,
  color_times = TRUE,
  show_insignificant_grey = FALSE,
  hide_insignificant = FALSE,
  fractional = F,
  theoretical = F
)
}
\arguments{
\item{p_dat}{data produced by the \code{\link{create_p_diff_uptake_dataset}}
function}

\item{state_1}{selected biological state for given protein}

\item{state_2}{selected biological state for given protein}

\item{adjust_axes}{\code{\link{logical}}, indicator if the X-axis 
is symmetrical in relation to 0}

\item{show_confidence_limits}{\code{\link{logical}}, indicates if the hybrid 
test confidence levels are shown}

\item{confidence_level}{confidence level for the test, from range [0, 1]. It
should be the same as used to prepare p_dat}

\item{color_times}{\code{\link{logical}}, indicator if different time points 
are distinguishable by color}

\item{show_insignificant_grey}{\code{\link{logical}}, indicator if the 
values not passing the test are shown in grey}

\item{hide_insignificant}{\code{\link{logical}}, indicator if the 
values not passing the test are hidden}

\item{fractional}{\code{logical}, indicator if values are shown 
in fractional form}

\item{theoretical}{\code{logical}, indicator if values are 
calculated using theoretical controls}
}
\value{
a \code{\link{ggplot}} object
}
\description{
Volcano plot for differential deuterim uptake
between two biological states
}
\details{
The function \code{\link{plot_volcano}} generates the 
volcano plot based on supplied p_dat. 
On X-axis there is differential deuterium uptake in selected form.
On Y-axis there is the P-value from t-Student test between two
biological states. Based on selected confidence level, the confidence
limits are calculated to indicate statistically significant values -
shown as red dotted lines. The values that are in upper left and right
corners pass the hybrid test.
}
\examples{
dat <-  read_hdx(system.file(package = "HaDeX", "HaDeX/data/KD_180110_CD160_HVEM.csv"))
p_dat <- create_p_diff_uptake_dataset(dat)
plot_volcano(p_dat, show_confidence_limits = T)

plot_volcano(p_dat, show_confidence_limits = T, show_insignificant_grey = T)
plot_volcano(p_dat, show_confidence_limits = T, hide_insignificant = T)
plot_volcano(p_dat, show_confidence_limits = T, hide_insignificant = T, show_insignificant_grey = T)

}
\references{
Hageman, T. S. & Weis, D. D. Reliable Identification of Significant
Differences in Differential Hydrogen Exchange-Mass Spectrometry Measurements
Using a Hybrid Significance Testing Approach. Anal Chem 91, 8008–8016 (2019).

Houde, D., Berkowitz, S.A., and Engen, J.R. (2011).
The Utility of Hydrogen/Deuterium Exchange Mass Spectrometry in
Biopharmaceutical Comparability Studies. J Pharm Sci 100, 2071–2086.
}
\seealso{
\code{\link{create_p_diff_uptake_dataset}}
}
