% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_coverage_heatmap_data.R
\name{show_coverage_heatmap_data}
\alias{show_coverage_heatmap_data}
\title{Coverage heatmap data}
\usage{
show_coverage_heatmap_data(x_dat, value = NULL)
}
\arguments{
\item{x_dat}{data created using calculate_ or create_ 
function}

\item{value}{value to be presented}
}
\description{
This function prepares the data used in 
coverage heatmap to be shown in user-friendly way.
}
\examples{
# auc data
auc_dat <- calculate_auc(create_uptake_dataset(dat))
show_coverage_heatmap_data(auc_dat, value = "auc")

# back-exchange
bex_dat <- calculate_back_exchange(dat, state = "CD160")
show_coverage_heatmap_data(bex_dat, value = "back_exchange")

}
