% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_data.R
\name{show_diff_uptake_data_confidence}
\alias{show_diff_uptake_data_confidence}
\title{Show differential uptake data with confidence levels}
\usage{
show_diff_uptake_data_confidence(
  diff_uptake_dat,
  theoretical = FALSE,
  fractional = FALSE,
  confidence_level = 0.98
)
}
\arguments{
\item{diff_uptake_dat}{data produced by \code{\link{create_diff_uptake_dataset}}
function.}

\item{theoretical}{\code{logical}, determines if values are theoretical.}

\item{fractional}{\code{logical}, determines if values are fractional.}

\item{confidence_level}{confidence level for the test, from range [0, 1].}
}
\value{
a \code{\link{data.frame}} object.
}
\description{
Show differential uptake data with confidence levels
}
\details{
This function subsets the dataset based on provided criteria,
rounds the numerical values (4 places) and changes the column names 
to user-friendly ones. 
This data is available in the GUI.
}
\examples{
dat <- read_hdx(system.file(package = "HaDeX", "HaDeX/data/KD_180110_CD160_HVEM.csv"))
diff_uptake_dat <- create_diff_uptake_dataset(dat)
show_diff_uptake_data_confidence(diff_uptake_dat)

}
\seealso{
\code{\link{create_diff_uptake_dataset}}
\code{\link{plot_differential}}
}
