% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicates.R
\name{show_replicate_histogram_data}
\alias{show_replicate_histogram_data}
\title{Show replicate data}
\usage{
show_replicate_histogram_data(rep_dat)
}
\arguments{
\item{rep_dat}{replicate data, created by 
\code{\link{create_replicate_dataset}} function.}
}
\value{
a \code{\link{data.frame}} object.
}
\description{
Show histogram data on number of replicates per
peptide in one or multiple time point of measurement.
}
\details{
The function shows the information about
number of replicates for peptides in one or multiple
time point of measurement, depends on supplied data.
}
\examples{
dat <- read_hdx(system.file(package = "HaDeX", "HaDeX/data/KD_180110_CD160_HVEM.csv"))
rep_dat <- create_replicate_dataset(dat)
show_replicate_histogram_data(rep_dat)

}
\seealso{
\code{\link{create_replicate_dataset}}
\code{\link{plot_replicate_histogram}}
}
