% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdx_data_class.R
\name{validate_hdx_data}
\alias{validate_hdx_data}
\title{Validator on hdx_data class}
\usage{
validate_hdx_data(hdx_data)
}
\arguments{
\item{hdx_data}{\code{\link{hdx_data}} object.}
}
\description{
Validator on the content of an hdx_data object.
}
