% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_stat_dependency.R
\name{add_stat_dependency}
\alias{add_stat_dependency}
\title{Calculates confidence limits}
\usage{
add_stat_dependency(
  calc_dat,
  confidence_limit = 0.98,
  theoretical = FALSE,
  fractional = TRUE
)
}
\arguments{
\item{calc_dat}{processed data from DynamX file - using \code{\link{prepare_dataset}}}

\item{confidence_limit}{confidence limit chosen by user - from range [0, 1].}

\item{theoretical}{\code{logical}, determines if values are theoretical}

\item{fractional}{\code{logical}, determines if values are fractional}
}
\value{
calc_dat extended by column specifying if given peptide is relevant in given confidence limit. 
The value of the confidence limit is added as an attribute - as well as parameters used to calculate 
(theoretical/fractional)
}
\description{
Returns relation with confidence limits for each peptide.
}
\details{
...
}
\examples{
#load example data
dat <- read_hdx(system.file(package = "HaDeX", 
                            "HaDeX/data/KD_180110_CD160_HVEM.csv"))
                            
# TODO

}
\seealso{
\code{\link{read_hdx}} \code{\link{prepare_dataset}}
}
