% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_stat_dependency.R
\name{add_stat_dependency}
\alias{add_stat_dependency}
\title{Calculates confidence limits}
\usage{
add_stat_dependency(
  calc_dat,
  confidence_level = 0.98,
  theoretical = FALSE,
  fractional = TRUE
)
}
\arguments{
\item{calc_dat}{data produced by \code{\link{calculate_diff_uptake}}
funcion.}

\item{confidence_level}{confidence limit - from range [0, 1].}

\item{theoretical}{\code{logical}, determines if values are theoretical.}

\item{fractional}{\code{logical}, determines if values are fractional.}
}
\value{
calc_dat extended by column specifying if given peptide is relevant in
given confidence limit. The value of the confidence limit is added as an attribute 
- as well as parameters used to calculate (theoretical/fractional).
}
\description{
Returns relation with confidence limits for each peptide.
}
\details{
This function checks if the values are statistically significant based 
on provided criteria using Houde test.
}
\examples{
dat <- read_hdx(system.file(package = "HaDeX", 
                            "HaDeX/data/KD_180110_CD160_HVEM.csv"))
calc_dat <- calculate_diff_uptake(dat)
result <- add_stat_dependency(calc_dat)
head(result)                            

}
\references{
Houde, D., Berkowitz, S.A., and Engen, J.R. (2011). 
The Utility of Hydrogen/Deuterium Exchange Mass Spectrometry in 
Biopharmaceutical Comparability Studies. J Pharm Sci 100, 2071–2086.
}
\seealso{
\code{\link{read_hdx}} 
\code{\link{calculate_diff_uptake}}
\code{\link{calculate_confidence_limit_values}}
}
