% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_confidence_limit_values.R
\name{calculate_confidence_limit_values}
\alias{calculate_confidence_limit_values}
\title{calculate_confidence_limit_values}
\usage{
calculate_confidence_limit_values(calc_dat, confidence_limit = 0.98,
  theoretical = FALSE, relative = TRUE)
}
\arguments{
\item{calc_dat}{processed data from DynamiX file - using prepare_dataset}

\item{confidence_limit}{confidence limit chosen by user - from range [0, 1]. default : 0.98}

\item{theoretical}{logical value to determine if plot is theoretical or not. default : false}

\item{relative}{logical value to determine if values are relative or absolute. default : true}
}
\value{
range of confidence limit interval
}
\description{
Calculates confidence limit values for prepared datased, based on chosen parameters.
}
\examples{
dat <- read_hdx(system.file(package = "HaDeX", "HaDeX/data/KD_180110_CD160_HVEM.csv"))
calc_dat <- prepare_dataset(dat,
                            in_state_first = "CD160_0.001",
                            chosen_state_first = "CD160_1",
                            out_state_first = "CD160_1440",
                            in_state_second = "CD160_HVEM_0.001",
                            chosen_state_second = "CD160_HVEM_1",
                            out_state_second = "CD160_HVEM_1440")   
calculate_confidence_limit_values(calc_dat = calc_dat,
                                  confidence_limit = 0.99,
                                  theoretical = FALSE, 
                                  relative = TRUE)                          
                              
}
