% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinetics.R
\name{calculate_kinetics}
\alias{calculate_kinetics}
\title{Calculate kinetics data for a peptide in specific state}
\usage{
calculate_kinetics(
  dat,
  protein = dat[["Protein"]][1],
  sequence,
  state,
  start,
  end,
  time_0,
  time_100,
  deut_part = 0.9
)
}
\arguments{
\item{dat}{dat data imported by the \code{\link{read_hdx}} function.}

\item{protein}{protein chosen protein.}

\item{sequence}{sequence of chosen peptide.}

\item{state}{biological state of chosen peptide.}

\item{start}{start position of chosen peptide.}

\item{end}{end position of chosen peptide.}

\item{time_0}{minimal exchange control time point of measurement.}

\item{time_100}{maximal exchange control time point of measurement.}

\item{deut_part}{deuterium percentage in solution used in experiment, 
value from range [0, 1].}
}
\value{

}
\description{
Calculate kinetics of the hydrogen-deuteration exchange 
for given peptide.
}
\details{
The function calculates deuteration data for all available data points 
for given peptide in chosen biological state.. 
All four variants (relative & theoretical combinations) of deuterium uptake computations 
are supported. Manual correction of percentage of deuterium the protein was exposed 
to during the exchange in theoretical calculations is provided.
To visualize obtained data we recommend using \code{\link{plot_kinetics}} function.
The first version doesn't support filled Modification and Fragment columns.
IMPORTANT! The kinetic data is often described as deuterium uptake curve data. 
We use this terms interchangeable.
}
\examples{
dat <- read_hdx(system.file(package = "HaDeX", 
                            "HaDeX/data/KD_180110_CD160_HVEM.csv"))

calculate_kinetics(dat, 
                   protein = "db_CD160",
                   sequence = "INITSSASQEGTRLN", 
                   state = "CD160",
                   start = 1, 
                   end = 15,
                   time_0 = 0.001, 
                   time_100 = 1440)
       
}
\seealso{
\code{\link{read_hdx}} 
\code{\link{calculate_state_uptake}} 
\code{\link{plot_kinetics}}
}
