% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_state_deuteration.R
\name{calculate_state_deuteration}
\alias{calculate_state_deuteration}
\title{calculate_state_deuteration}
\usage{
calculate_state_deuteration(dat, protein, state, time_in, time_chosen,
  time_out)
}
\arguments{
\item{dat}{dat}

\item{protein}{protein}

\item{state}{state}

\item{time_in}{time in}

\item{time_chosen}{time chosen}

\item{time_out}{time out}
}
\value{
data.frame
}
\description{
calculate_state_deuteration
}
\examples{
dat <- read_hdx(system.file(package = "HaDeX", "HaDeX/data/KD_180110_CD160_HVEM.csv"))

}
