% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparison_plot.R
\name{comparison_plot}
\alias{comparison_plot}
\title{comparison_plot}
\usage{
comparison_plot(calc_dat, x_lab = NA, y_lab = NA, plot_title = NA,
  theoretical = FALSE, state_first = "state_first",
  state_second = "state_second")
}
\arguments{
\item{calc_dat}{processed data from DynamiX file - using prepare_dataset}

\item{x_lab}{x label chosen by user. If unused, default label is chosen}

\item{y_lab}{y label chosen by user. If unused, default label is chosen}

\item{plot_title}{title label chosen by user. If unused, default label is chosen}

\item{theoretical}{logical value to determine if plot is theoretical or not. default : false}

\item{state_first}{first state name}

\item{state_second}{second state name}
}
\description{
Produces comparison_plot based on previously processed data - theoretical or experimental. User can change labels if needed.
}
\examples{
dat <- read_hdx(system.file(package = "HaDeX", "HaDeX/data/KD_180110_CD160_HVEM.csv"))
calc_dat <- prepare_dataset(dat,
                            in_state_first = "CD160_0.001",
                            chosen_state_first = "CD160_1",
                            out_state_first = "CD160_1440",
                            in_state_second = "CD160_HVEM_0.001",
                            chosen_state_second = "CD160_HVEM_1",
                            out_state_second = "CD160_HVEM_1440")                             
comparison_plot(calc_dat = calc_dat,
                theoretical = TRUE,
                state_first = "CD160",
                state_second = "CD160_HVEM")
comparison_plot(calc_dat = calc_dat,
                state_first = "CD160",
                state_second = "CD160_HVEM")                

}
