% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparison_plot.R
\name{comparison_plot}
\alias{comparison_plot}
\title{comparison_plot}
\usage{
comparison_plot(calc_dat, x_lab = NA, y_lab = NA, plot_title = NA,
  theoretical = FALSE)
}
\arguments{
\item{calc_dat}{processed data from DynamiX file - using prepare_dataset}

\item{x_lab}{x label chosen by user. If unused, default label is chosen}

\item{y_lab}{y label chosen by user. If unused, default label is chosen}

\item{plot_title}{title label chosen by user. If unused, default label is chosen}

\item{theoretical}{logical value to determine if plot is theoretical or not. default : false}
}
\description{
Produces comparison_plot based on previously processed data - theoretical or experimental. User can change labels if needed.
}
