% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_control.R
\name{create_quality_control_dataset}
\alias{create_quality_control_dataset}
\title{Experiment quality control}
\usage{
create_quality_control_dataset(
  dat,
  protein,
  state_1,
  state_2,
  time_t,
  time_0,
  deut_part = 0.9
)
}
\arguments{
\item{dat}{data read by \code{\link{read_hdx}}}

\item{protein}{...}

\item{state_1}{state of the first peptide}

\item{state_2}{state of the second peptide}

\item{time_t}{chosen time point}

\item{time_0}{`in` time}

\item{deut_part}{...}
}
\value{
\code{\link{data.frame}} object with mean uncertainty per different 
`out_time` value. 
The values are shown as percentages.
}
\description{
Checks how the uncertainty changes in a function of `out_time`.
}
\details{
The function calculates mean uncertainty of all peptides and its 
uncertainty (standard error) based on given `in_time` and `chosen_time` 
as a function of `out_time`. Both theoretical and experimental results for 
each state and their difference are supplied for comparison but only 
experimental calculations depends on `out_time` variable. The results are 
either in form of fractional or absolute values depending on the `fractional` 
parameter supplied by the user. 
This data can be useful for general overview of the experiment and analyse 
of the chosen time parameters.
}
\examples{
dat <- read_hdx(system.file(package = "HaDeX", "HaDeX/data/KD_180110_CD160_HVEM.csv"))

(qc_dat <- create_quality_control_dataset(dat = dat,
                                          protein = "db_CD160",
                                          state_1 = "CD160",
                                          state_2 = "CD160_HVEM", 
                                          time_t = 1, 
                                          time_0 = 0.001,
                                          deut_part = 0.9))    
                           
plot_quality_control(qc_dat)

}
\seealso{
\code{\link{read_hdx}}
\code{\link{calculate_state_uptake}}
\code{\link{plot_quality_control}}
\code{\link{show_quality_control_data}}
}
