% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_butterfly_differential_elements.R
\name{generate_butterfly_differential_plot}
\alias{generate_butterfly_differential_plot}
\title{generate_butterfly_differential_plot}
\usage{
generate_butterfly_differential_plot(
  butterfly_diff_dat,
  theoretical = FALSE,
  fractional = FALSE
)
}
\arguments{
\item{butterfly_diff_dat}{...}

\item{theoretical}{\code{logical}, determines if values are theoretical}

\item{fractional}{\code{logical}, determines if values are fractional}
}
\value{

}
\description{
generate_butterfly_differential_plot
}
\details{
This plot is visible in GUI.
}
\seealso{

}
