% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_butterfly_elements.R
\name{generate_butterfly_plot}
\alias{generate_butterfly_plot}
\title{generate_butterfly_plot}
\usage{
generate_butterfly_plot(
  butterfly_dat,
  theoretical = FALSE,
  fractional = FALSE,
  uncertainty_type = c("ribbon", "bars")
)
}
\arguments{
\item{butterfly_dat}{...}

\item{theoretical}{\code{logical}, determines if values are theoretical}

\item{fractional}{\code{logical}, determines if values are fractional}

\item{uncertainty_type}{...}
}
\value{
...
}
\description{
Generates butterfly plot based on supplied data
and parameters.
}
\details{
This plot is visible in GUI.
}
\seealso{
...
}
