% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_butterfly_elements.R
\name{generate_butterfly_plot}
\alias{generate_butterfly_plot}
\title{generate_butterfly_plot}
\usage{
generate_butterfly_plot(dat, theoretical, fractional)
}
\arguments{
\item{dat}{data as imported by the \code{\link{read_hdx}} function}

\item{theoretical}{\code{logical}, determines if values are theoretical}

\item{fractional}{\code{logical}, determines if values are fractional}
}
\value{
...
}
\description{
Generates butterfly plot based on supplied data
and parameters.
}
\details{
This plot is visible in GUI.
}
\seealso{
...
}
