% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_chiclet_elements.R
\name{generate_chiclet_data}
\alias{generate_chiclet_data}
\title{generate_chiclet_data}
\usage{
generate_chiclet_data(chiclet_dat, theoretical = FALSE, fractional = FALSE)
}
\arguments{
\item{chiclet_dat}{...}

\item{theoretical}{\code{logical}, determines if values are theoretical}

\item{fractional}{\code{logical}, determines if values are fractional}
}
\value{
...
}
\description{
Generates chiclet data, based on the supplied
parameters.
}
\details{
This data is available in the GUI. 
All of the numerical values are rounded to 4 places after the dot!!
}
\seealso{
...
}
