% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_chiclet_differential_elements.R
\name{generate_chiclet_differential_data}
\alias{generate_chiclet_differential_data}
\title{generate_chiclet_differential_data}
\usage{
generate_chiclet_differential_data(
  chiclet_diff_dat,
  theoretical = FALSE,
  fractional = FALSE
)
}
\arguments{
\item{chiclet_diff_dat}{...}

\item{theoretical}{\code{logical}, determines if values are theoretical}

\item{fractional}{\code{logical}, determines if values are fractional}
}
\value{

}
\description{
generate_chiclet_differential_data
}
\details{

}
\seealso{

}
