% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_chiclet_differential_elements.R
\name{generate_chiclet_differential_plot}
\alias{generate_chiclet_differential_plot}
\title{generate_chiclet_differential_plot}
\usage{
generate_chiclet_differential_plot(
  chiclet_diff_dat,
  theoretical = FALSE,
  fractional = FALSE,
  show_uncertainty = "ribbon"
)
}
\arguments{
\item{chiclet_diff_dat}{...}

\item{theoretical}{\code{logical}, determines if values are theoretical}

\item{fractional}{\code{logical}, determines if values are fractional}

\item{show_uncertainty}{...}
}
\value{

}
\description{
generate_chiclet_differential_plot
}
\details{
This plot is visible in GUI.
}
\seealso{

}
