% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_comparison_data.R
\name{generate_comparison_data}
\alias{generate_comparison_data}
\title{generate_comparison_data}
\usage{
generate_comparison_data(dat, theoretical, relative, protein)
}
\arguments{
\item{dat}{custom data format, produced by \code{\link{calculate_state_deuteration}}}

\item{theoretical}{...}

\item{relative}{...}
}
\value{
...
}
\description{
Generates comparison data, based on the supplied
parameters.
}
\details{
This data is available in the GUI. The names of the parameters
and variables will be changed later after the glossary project.
}
\seealso{
...
}
