% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_comparison_elements.R
\name{generate_comparison_data}
\alias{generate_comparison_data}
\title{generate_comparison_data}
\usage{
generate_comparison_data(dat, theoretical, fractional, protein)
}
\arguments{
\item{dat}{custom data format, produced by \code{\link{calculate_state_deuteration}}}

\item{theoretical}{\code{logical}, determines if values are theoretical}

\item{fractional}{\code{logical}, determines if values are fractional}
}
\value{
...
}
\description{
Generates comparison data, based on the supplied
parameters.
}
\details{
This data is available in the GUI. 
All of the numerical values are rounded to 4 places after the dot!!
}
\seealso{
...
}
