% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_comparison_elements.R
\name{generate_comparison_data}
\alias{generate_comparison_data}
\title{Generate comparison data}
\usage{
generate_comparison_data(dat, theoretical = FALSE, fractional = FALSE, protein)
}
\arguments{
\item{dat}{data produced by \code{\link{calculate_state_deuteration}} or
\code{\link{generate_comparison_dataset}} function.}

\item{theoretical}{\code{logical}, determines if values are theoretical.}

\item{fractional}{\code{logical}, determines if values are fractional}

\item{protein}{chosen protein.}
}
\value{
a \code{data.frame} object.
}
\description{
Generate comparison data
}
\details{
This function subsets the dataset based on provided criteria,
rounds the numerical values (4 places) and changes the column names 
to user-friendly ones. 
This data is available in the GUI.
}
\examples{
dat <- read_hdx(system.file(package = "HaDeX", "HaDeX/data/KD_180110_CD160_HVEM.csv"))
comparison_dat <- generate_comparison_dataset(dat)
generate_comparison_data(comparison_dat)

}
\seealso{
\code{\link{calculate_state_deuteration}} 
\code{\link{generate_comparison_dataset}}
\code{\link{generate_comparison_plot}}
}
