% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_comparison_elements.R
\name{generate_comparison_dataset}
\alias{generate_comparison_dataset}
\title{Generates comparison dataset}
\usage{
generate_comparison_dataset(
  dat,
  protein = unique(dat[["Protein"]])[1],
  states = unique(dat[["State"]]),
  time_0 = 0.001,
  time_t = 1,
  time_100 = 1440,
  deut_part = 1
)
}
\arguments{
\item{dat}{data imported by the \code{\link{read_hdx}} function.}

\item{protein}{chosen protein.}

\item{states}{vector of states (for chosen protein), for which the 
calculations are done.}

\item{time_0}{minimal exchange control time point of measurement.}

\item{time_t}{time point of the measurement for which the calculations
are done.}

\item{time_100}{maximal exchange control time point of measurement.}

\item{deut_part}{deuterium percentage in solution used in experiment, 
value from range [0, 1].}
}
\value{
a data frame with calculated deuterium uptake, fractional deuterium 
uptake, theoretical deuterium uptake, theoretical fractional deuterium uptake
and their uncertainty, based on supplied parameters.
}
\description{
Generates comparison dataset
}
\details{
Function \code{\link{generate_comparison_dataset}} is a 
wrapper for \code{\link{calculate_state_deuteration}} function, calls 
this function for all states in states vector.
}
\examples{
dat <- read_hdx(system.file(package = "HaDeX", "HaDeX/data/KD_180110_CD160_HVEM.csv"))
comparison_dat <- generate_comparison_dataset(dat)
head(comparison_dat)

}
\seealso{
\code{\link{calculate_state_deuteration}}
\code{\link{generate_comparison_plot}}
\code{\link{generate_comparison_data}}
}
