% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_comparison_elements.R
\name{generate_comparison_dataset}
\alias{generate_comparison_dataset}
\title{generate_comparison_dataset}
\usage{
generate_comparison_dataset(
  dat,
  protein = unique(dat[["Protein"]])[1],
  states = unique(dat[["State"]]),
  time_0 = 0.001,
  time_t = 1,
  time_100 = 1440,
  deut_part = 1
)
}
\arguments{
\item{dat}{...}

\item{protein}{...}

\item{states}{vector of two states (for chosen protein). The deuterium 
uptake difference is calculates as first state - second state, so the 
order of states matter.}

\item{time_0}{minimal exchange control}

\item{time_t}{maximal exchange control}

\item{time_100}{...}

\item{deut_part}{...}
}
\value{
..
}
\description{
Generates data set for given states
}
\details{
This is internal function, not used in the application.
This function is a wrapper for \code{\link{calculate_state_deuteration}}.
}
\seealso{
...
}
