% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_comparison_elements.R
\name{generate_comparison_plot}
\alias{generate_comparison_plot}
\title{generate_comparison_plot}
\usage{
generate_comparison_plot(dat, theoretical = FALSE, fractional = FALSE)
}
\arguments{
\item{dat}{produced by \code{\link{calculate_state_deuteration}} function}

\item{theoretical}{\code{logical}, determines if values are theoretical}

\item{fractional}{\code{logical}, determines if values are fractional}
}
\value{
...
}
\description{
Generates comparison plot based on supplied data
and parameters.
}
\details{
This plot is visible in GUI.
}
\seealso{
...
}
