% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_differential_data.R
\name{generate_differential_data}
\alias{generate_differential_data}
\title{generate_differential_data}
\usage{
generate_differential_data(
  dat,
  theoretical,
  relative,
  confidence_limit_1,
  confidence_limit_2
)
}
\arguments{
\item{dat}{custom format, produced by 
\code{\link{generate_differential_data_set}}}

\item{theoretical}{...}

\item{relative}{...}

\item{confidence_limit_1}{...}

\item{confidence_limit_2}{...}
}
\value{
...
}
\description{
Generates differential data, based on the supplied
parameters.
}
\details{
This data is available in the GUI. 
All of the numerical values are rounded to 4 places after the dot!!
}
\seealso{
...
}
