% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_differential_elements.R
\name{generate_differential_data_set}
\alias{generate_differential_data_set}
\title{generate_differential_data_set}
\usage{
generate_differential_data_set(
  dat,
  states = unique(dat[["State"]])[1:2],
  protein = unique(dat[["Protein"]][1]),
  time_0 = 0.001,
  time_t = 1,
  time_100 = 1440,
  deut_part = 1
)
}
\arguments{
\item{dat}{...}

\item{states}{vector of two states to calculate difference between them, 
the order is important (state first - state second)}

\item{protein}{...}

\item{time_0}{...}

\item{time_t}{...}

\item{time_100}{...}

\item{deut_part}{...}
}
\value{
...
}
\description{
Generate the data frame with differential data from two
provided states - experimental/fractional calculations with the
uncertainty, based on supplied parameters.
}
\details{
The names of the parameters and variables will be changed 
later after the glossary project.
}
\examples{
# load example data
dat <- read_hdx(system.file(package = "HaDeX", "HaDeX/data/KD_180110_CD160_HVEM.csv"))

# calculate differential values between states "CD160" and "CD160_HVEM" for protein "db_CD160"
generate_differential_data_set(dat = dat, states = c("CD160", "CD160_HVEM"), protein = "db_CD160", 
                               time_0 = 0.001, time_t = 5.000, time_100 = 1440.000)

}
\seealso{
...
}
