% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_differential_elements.R
\name{generate_differential_plot}
\alias{generate_differential_plot}
\title{generate_differential_plot}
\usage{
generate_differential_plot(
  dat,
  theoretical,
  fractional,
  confidence_limit,
  confidence_limit_2
)
}
\arguments{
\item{dat}{produced by \code{\link{generate_differential_data_set}} function}

\item{theoretical}{\code{logical}, determines if values are theoretical}

\item{fractional}{\code{logical}, determines if values are fractional}

\item{confidence_limit}{...}

\item{confidence_limit_2}{...}
}
\value{
...
}
\description{
Generates differential (Woods) plot with confidence values
based on supplied data and parameters.
}
\details{
This plot is visible in GUI.
}
\seealso{
...
}
