% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_kinetic_data.R
\name{generate_kinetic_data}
\alias{generate_kinetic_data}
\title{generate_kinetic_data}
\usage{
generate_kinetic_data(dat, theoretical, relative)
}
\arguments{
\item{dat}{custom format, produced by 
\code{\link{generate_kinetic_data_set}}}

\item{theoretical}{...}

\item{relative}{...}
}
\value{
...
}
\description{
Generates deuterium uptake data, based on the supplied
parameters.
}
\details{
This data is available in the GUI. The names of the parameters
and variables will be changed later after the glossary project.
}
\seealso{
...
}
