% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_kinetic_data_set.R
\name{generate_kinetic_data_set}
\alias{generate_kinetic_data_set}
\title{generate_kinetic_data_set}
\usage{
generate_kinetic_data_set(
  dat,
  peptide_list,
  protein,
  time_in,
  time_out,
  deut_concentration
)
}
\arguments{
\item{dat}{...}

\item{peptide_list}{...}

\item{protein}{...}

\item{time_in}{...}

\item{time_out}{...}

\item{deut_concentration}{...}
}
\value{
...
}
\description{
Generates the data set of deuterium uptake between selected 
time points based on supplied peptide list.
}
\details{
This is a wrapper for \code{\link{calculate_kinetics}}, but for
the peptide list instead of one peptide. The names of the parameters and 
variables will be changed later after the glossary project.
}
\seealso{
...
}
