% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_overlap_elements.R
\name{generate_overlap_distribution_data}
\alias{generate_overlap_distribution_data}
\title{generate_overlap_distribution_data}
\usage{
generate_overlap_distribution_data(
  dat,
  protein,
  state,
  start,
  end,
  protein_sequence
)
}
\arguments{
\item{dat}{...}

\item{protein}{...}

\item{state}{...}

\item{start}{...}

\item{end}{...}

\item{protein_sequence}{...}
}
\value{
...
}
\description{
Generates the data set of frequency of overlap of
each amino in the protein sequence.
}
\details{
This data is available in the GUI.
}
\seealso{
...
}
