% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_overlap_distribution_plot.R
\name{generate_overlap_distribution_plot}
\alias{generate_overlap_distribution_plot}
\title{generate_overlap_distribution_plot}
\usage{
generate_overlap_distribution_plot(dat, start, end)
}
\arguments{
\item{dat}{produced by \code{\link{generate_overlap_distribution_plot}}
function}

\item{start}{...}

\item{end}{...}
}
\value{
...
}
\description{
Generates overlap distribution plot based on supplied data
and parameters.
}
\details{
This plot is visible in GUI.
}
\seealso{
...
}
