% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_overlap_plot.R
\name{generate_overlap_plot}
\alias{generate_overlap_plot}
\title{generate_overlap_plot}
\usage{
generate_overlap_plot(dat)
}
\arguments{
\item{dat}{produced by \code{\link{generate_overlap_plot}}
function}
}
\value{
...
}
\description{
Generates overapping peptide plot based on supplied data
and parameters.
}
\details{
This plot is visible in GUI.
}
\seealso{
...
}
