% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_quality_control_elements.R
\name{generate_quality_control_data}
\alias{generate_quality_control_data}
\title{generate_quality_control_data}
\usage{
generate_quality_control_data(dat)
}
\arguments{
\item{dat}{custom format, produced by \code{\link{quality_control}}, 
scaled if necessary.}
}
\value{
...
}
\description{
Generates quality control data, based on the supplied
parameters.
}
\details{
This data is available in the GUI. The names of the parameters
and variables will be changed later after the glossary project.
}
\seealso{
...
}
