% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_quality_control_elements.R
\name{generate_quality_control_plot}
\alias{generate_quality_control_plot}
\title{generate_quality_control_plot}
\usage{
generate_quality_control_plot(dat)
}
\arguments{
\item{dat}{produced by \code{\link{quality_control}} function, 
scaled if necessary.}
}
\value{
...
}
\description{
Generates quality control plot based on supplied data.
}
\details{
This plot is visible in GUI.
}
\seealso{
...
}
