% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_volcano_elements.R
\name{generate_volcano_data}
\alias{generate_volcano_data}
\title{generate_volcano_data}
\usage{
generate_volcano_data(vol_data)
}
\arguments{
\item{vol_data}{...}
}
\description{
generate_volcano_data
}
