% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_volcano_elements.R
\name{generate_volcano_plot}
\alias{generate_volcano_plot}
\title{generate_volcano_plot}
\usage{
generate_volcano_plot(
  vol_data,
  state_1 = "",
  state_2 = "",
  adjust_axes = TRUE,
  show_confidence_limits = FALSE,
  confidence_level = 0.98
)
}
\arguments{
\item{vol_data}{...}

\item{state_1}{...}

\item{state_2}{...}

\item{adjust_axes}{...}

\item{show_confidence_limits}{...}

\item{confidence_level}{...}
}
\description{
Generates volcano plot based on supplied volcano data
}
