% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_volcano_elements.R
\name{generate_volcano_plot}
\alias{generate_volcano_plot}
\title{generate_volcano_plot}
\usage{
generate_volcano_plot(vol_data, state_1 = "", state_2 = "")
}
\arguments{
\item{vol_data}{...}
}
\description{
Generates volcano plot based on supplied volcano data
}
