% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HaDeX_gui.R
\name{HaDeX_gui}
\alias{HaDeX_gui}
\title{HaDeX Graphical User Interface}
\usage{
HaDeX_gui(port = getOption("shiny.port"))
}
\arguments{
\item{port}{The TCP port. See \code{\link[shiny]{runApp}}.}
}
\description{
Launches graphical user interface.
}
\section{Warning }{
 Any ad-blocking software may cause malfunctions.
}

\examples{
HaDeX_gui()

}
