% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HaDeXify.R
\name{HaDeXify}
\alias{HaDeXify}
\title{HaDeX customized ggplot theme}
\usage{
HaDeXify(plt)
}
\arguments{
\item{plt}{ggplot object. Plot to HaDeXify.}
}
\description{
This function HaDeXifies plot. It adds HaDeX logo and ggplot theme.
}
