% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amino_distribution.R
\name{plot_amino_distribution}
\alias{plot_amino_distribution}
\title{generate_amino_distribution}
\usage{
plot_amino_distribution(
  position_in_sequence,
  hydro_properties,
  protein,
  charge_colors
)
}
\arguments{
\item{position_in_sequence}{custom format}

\item{hydro_properties}{...}

\item{protein}{...}

\item{charge_colors}{...}
}
\value{
...
}
\description{
Generates amino distribution based on the 
protein sequence and shows if the amino acid is hydrophobic
or hydrophylic.
}
\details{
The data for this function is not packaged yet.
}
\seealso{
...
}
