% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_coverage_heatmap.R
\name{plot_coverage_heatmap}
\alias{plot_coverage_heatmap}
\title{Coverage heatmap}
\usage{
plot_coverage_heatmap(
  x_dat,
  protein = x_dat[["Protein"]][1],
  state = NULL,
  value = NULL,
  time_t = NULL
)
}
\arguments{
\item{x_dat}{data created using calculate_ or create_ 
function}

\item{protein}{selected protein}

\item{value}{value to be presented}

\item{time_t}{...}
}
\value{
a \code{\link{ggplot}} object
}
\description{
Coverage heatmap with color indicating
specific value
}
\details{
...
}
\examples{
dat <- read_hdx(system.file(package = "HaDeX", "HaDeX/data/KD_180110_CD160_HVEM.csv"))
x_dat <- create_uptake_dataset(dat, states = "CD160")
plot_coverage_heatmap(x_dat)
plot_coverage_heatmap(x_dat, value = "frac_deut_uptake", time_t = 0.167)
plot_coverage_heatmap(x_dat, value = "err_frac_deut_uptake", time_t = 0.167)

x_dat <- create_diff_uptake_dataset(dat)
plot_coverage_heatmap(x_dat)
plot_coverage_heatmap(x_dat, value = "diff_frac_deut_uptake")
plot_coverage_heatmap(x_dat, value = "diff_frac_deut_uptake", time_t = 0.167)
plot_coverage_heatmap(x_dat, value = "err_diff_frac_deut_uptake", time_t = 0.167)

}
\seealso{
\code{\link{read_hdx}}
\code{\link{plot_coverage}}
}
