% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_plots.R
\name{plot_differential}
\alias{plot_differential}
\title{Differential plot}
\usage{
plot_differential(
  dat,
  theoretical = FALSE,
  fractional = FALSE,
  confidence_limit = 0.98,
  confidence_limit_2 = 0.99
)
}
\arguments{
\item{dat}{produced by \code{\link{generate_differential_data_set}} function}

\item{theoretical}{\code{logical}, determines if values are theoretical}

\item{fractional}{\code{logical}, determines if values are fractional}

\item{confidence_limit}{...}

\item{confidence_limit_2}{...}
}
\value{
a \code{\link{ggplot}} object.
}
\description{
Differential plot
}
\details{
Function \code{\link{generate_differential_plot}} presents
provided data in a form of differential (Woods) plot. The plot show 
difference in exchange for two biological states, selected in 
generation of dataset at one time point of measurement .On X-axis 
there is a position in a sequence, with length of a segment of each 
peptide representing its length. On Y-axis there 
is deuterium uptake difference in chosen form. Error bars represents 
the combined and propagated uncertainty. 
This plot is visible in GUI.
}
\examples{
dat <- read_hdx(system.file(package = "HaDeX", "HaDeX/data/KD_180110_CD160_HVEM.csv"))
diff_uptake_dat <- generate_differential_data_set(dat)
plot_differential(diff_uptake_dat)

}
\seealso{
\code{\link{generate_differential_data_set}}
\code{\link{generate_differential_data}}
}
