% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_differential.R
\name{plot_differential}
\alias{plot_differential}
\title{Differential plot}
\usage{
plot_differential(
  diff_uptake_dat = NULL,
  diff_p_uptake_dat = NULL,
  time_t = NULL,
  theoretical = FALSE,
  fractional = FALSE,
  show_houde_interval = FALSE,
  show_tstud_confidence = FALSE,
  confidence_level = 0.98,
  all_times = FALSE,
  line_size = 1.5
)
}
\arguments{
\item{diff_uptake_dat}{produced by \code{\link{create_diff_uptake_dataset}} function.}

\item{diff_p_uptake_dat}{...}

\item{time_t}{...}

\item{theoretical}{\code{logical}, determines if values are theoretical.}

\item{fractional}{\code{logical}, determines if values are fractional.}

\item{show_houde_interval}{\code{logical}, determines if houde interval is shown.}

\item{show_tstud_confidence}{\code{logical}, determines if t-Student test validity 
is shown.}

\item{confidence_level}{confidence level for the test, from range [0, 1].}

\item{all_times}{...}

\item{line_size}{...}
}
\value{
a \code{\link{ggplot}} object.
}
\description{
Differential plot
}
\details{
Function \code{\link{plot_differential}} presents
provided data in a form of differential (Woods) plot. The plot shows
difference in exchange for two biological states, selected in
generation of dataset at one time point of measurement .On X-axis
there is a position in a sequence, with length of a segment of each
peptide representing its length. On Y-axis there
is deuterium uptake difference in chosen form. Error bars represents
the combined and propagated uncertainty.
The confidence limits based on provided confidence levels are shown
on the plot. The statistically significant values are in color (red if the 
difference is positive, blue if negative), and the insignificant values are 
grey. There are two confidence limits for comparison of the results, but
there is possibility to plot only one confidence limit, if the confidence
levels are the same.
This plot is visible in GUI.
}
\examples{
dat <- read_hdx(system.file(package = "HaDeX", "HaDeX/data/KD_180110_CD160_HVEM.csv"))
diff_uptake_dat <- create_diff_uptake_dataset(dat)
plot_differential(diff_uptake_dat = diff_uptake_dat, time_t = 0.167) # legend should be hidden
plot_differential(diff_uptake_dat = diff_uptake_dat, time_t = 0.167, line_size = 1) # legend should be hidden
plot_differential(diff_uptake_dat = diff_uptake_dat, all_times = T)
plot_differential(diff_uptake_dat = diff_uptake_dat, all_times = T, show_houde_interval = T)

diff_p_uptake_dat <- create_p_diff_uptake_dataset(dat)
plot_differential(diff_p_uptake_dat = diff_p_uptake_dat, all_times = T, show_tstud_confidence = T)
plot_differential(diff_p_uptake_dat = diff_p_uptake_dat, all_times = T, show_tstud_confidence = T, show_houde_interval = T)
plot_differential(diff_p_uptake_dat = diff_p_uptake_dat, show_tstud_confidence = T, show_houde_interval = T, all_times = F)

}
\seealso{
\code{\link{create_diff_uptake_dataset}}
\code{\link{show_diff_uptake_data}}
}
