% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinetics.R
\name{plot_differential_uptake_curve}
\alias{plot_differential_uptake_curve}
\title{Plot differential uptake curve}
\usage{
plot_differential_uptake_curve(
  diff_uptake_dat,
  sequence = NULL,
  theoretical = FALSE,
  fractional = FALSE,
  uncertainty_type = "ribbon",
  log_x = TRUE
)
}
\arguments{
\item{diff_uptake_dat}{produced by \code{\link{create_diff_uptake_dataset}} function}

\item{sequence}{}

\item{theoretical}{\code{logical}, determines if plot shows theoretical values.}

\item{fractional}{\code{logical}, determines if plot shows fractional values.}

\item{uncertainty_type}{type of presenting uncertainty, possible values:
"ribbon", "bars" or "bars + line".}

\item{log_x}{\code{logical}, determines if x axis shows logarithmic values.}
}
\value{
a \code{\link{ggplot2}} object.
}
\description{

}
\details{
Currently there is no possibility to plot multiple peptides on the plot.
}
\examples{
dat <- read_hdx(system.file(package = "HaDeX", "HaDeX/data/KD_180110_CD160_HVEM.csv"))
diff_uptake_dat <- create_diff_uptake_dataset(dat)
plot_differential_uptake_curve(diff_uptake_dat, sequence = "LCKDRSGDCSPETSLKQL")

}
\seealso{
\code{\link{read_hdx}}
\code{\link{create_diff_uptake_dataset}}
}
