% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_differential_uptake_curve.R
\name{plot_differential_uptake_curve}
\alias{plot_differential_uptake_curve}
\title{Plot differential uptake curve}
\usage{
plot_differential_uptake_curve(
  diff_uptake_dat = NULL,
  diff_p_uptake_dat = NULL,
  sequence = NULL,
  theoretical = FALSE,
  fractional = FALSE,
  uncertainty_type = "ribbon",
  log_x = TRUE,
  show_houde_interval = FALSE,
  show_tstud_confidence = FALSE
)
}
\arguments{
\item{diff_uptake_dat}{produced by \code{\link{create_diff_uptake_dataset}} function}

\item{sequence}{}

\item{theoretical}{\code{logical}, determines if plot shows theoretical values.}

\item{fractional}{\code{logical}, determines if plot shows fractional values.}

\item{uncertainty_type}{type of presenting uncertainty, possible values:
"ribbon", "bars" or "bars + line".}

\item{log_x}{\code{logical}, determines if x axis shows logarithmic values.}

\item{show_houde_interval}{\code{logical}, determines if houde interval is shown.}

\item{show_tstud_confidence}{\code{logical}, determines if t-Student test validity 
is shown.}
}
\value{
a \code{\link{ggplot2}} object.
}
\description{

}
\details{
Currently there is no possibility to plot multiple peptides on the plot.
}
\examples{
dat <- read_hdx(system.file(package = "HaDeX", "HaDeX/data/KD_180110_CD160_HVEM.csv"))
diff_uptake_dat <- create_diff_uptake_dataset(dat)
plot_differential_uptake_curve(diff_uptake_dat = diff_uptake_dat, sequence = "LCKDRSGDCSPETSLKQL")

diff_p_uptake_dat <- create_p_diff_uptake_dataset(dat)
plot_differential_uptake_curve(diff_p_uptake_dat = diff_p_uptake_dat, sequence = "LCKDRSGDCSPETSLKQL", show_houde_interval = TRUE)
plot_differential_uptake_curve(diff_p_uptake_dat = diff_p_uptake_dat, sequence = "LCKDRSGDCSPETSLKQL", show_houde_interval = TRUE, show_tstud_confidence = TRUE)
plot_differential_uptake_curve(diff_p_uptake_dat = diff_p_uptake_dat, sequence = "LCKDRSGDCSPETSLKQL", show_tstud_confidence = TRUE)

}
\seealso{
\code{\link{read_hdx}}
\code{\link{create_diff_uptake_dataset}}
}
