% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_plots.R
\name{plot_manhattan}
\alias{plot_manhattan}
\title{Manhattan plot}
\usage{
plot_manhattan(
  p_dat,
  plot_title = NULL,
  separate_times = T,
  confidence_level = NULL,
  show_confidence_limit = T
)
}
\arguments{
\item{p_dat}{data produced by the \code{\link{create_p_diff_uptake_dataset}}
function.}

\item{plot_title}{title for the plot. If not provided, it is constructed in a form:
"Differences between state_1 and state_2"}

\item{separate_times}{...}

\item{confidence_level}{confidence level for the test, from range [0, 1].}

\item{show_confidence_limit}{logical, indicates if the hybrid testing
confidence intervals are shown.}
}
\value{
a \code{\link{ggplot}} object.
}
\description{
Manhattan plot
}
\details{
...
}
\examples{
dat <-  read_hdx(system.file(package = "HaDeX", "HaDeX/data/KD_180110_CD160_HVEM.csv"))
p_dat <- create_p_diff_uptake_dataset(dat)
manhattan_plot(p_dat)

manhattan_plot(p_dat, separate_times = F)

manhattan_plot(p_dat, separate_times = F, show_confidence_limit = F)
 
}
\references{
Hageman, T. S. & Weis, D. D. Reliable Identification of Significant
Differences in Differential Hydrogen Exchange-Mass Spectrometry Measurements
Using a Hybrid Significance Testing Approach. Anal Chem 91, 8008–8016 (2019).
}
\seealso{
\code{\link{create_p_diff_uptake_dataset}}
}
