% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_plots.R
\name{plot_manhattan}
\alias{plot_manhattan}
\title{Manhattan plot}
\usage{
plot_manhattan(
  p_dat,
  plot_title = NULL,
  separate_times = T,
  times = NULL,
  confidence_level = NULL,
  show_confidence_limit = T,
  show_peptide_position = F
)
}
\arguments{
\item{p_dat}{data produced by the \code{\link{create_p_diff_uptake_dataset}}
function.}

\item{plot_title}{title for the plot. If not provided, it is constructed in a form:
"Differences between state_1 and state_2"}

\item{separate_times}{\code{logical}, indicates if the data should be seen on the same plot, 
or on separate plots for each time point of measurement.}

\item{times}{vector of time points of measurements to be included in the plot.}

\item{confidence_level}{confidence level for the test, from range [0, 1].}

\item{show_confidence_limit}{logical, indicates if the hybrid testing
confidence intervals are shown.}

\item{show_peptide_position}{\code{logical}, indicates if the peptide length
and position in  the  sequence is shown. Otherwise, the peptides are represented by their ID.}
}
\value{
a \code{\link{ggplot}} object.
}
\description{
Manhattan plot
}
\details{
...
}
\examples{
dat <-  read_hdx(system.file(package = "HaDeX", "HaDeX/data/KD_180110_CD160_HVEM.csv"))
p_dat <- create_p_diff_uptake_dataset(dat)
plot_manhattan(p_dat)

plot_manhattan(p_dat, separate_times = F)

plot_manhattan(p_dat, show_peptide_position = T, separate_times = F)

plot_manhattan(p_dat, separate_times = F, show_confidence_limit = F)
 
}
\references{
Hageman, T. S. & Weis, D. D. Reliable Identification of Significant
Differences in Differential Hydrogen Exchange-Mass Spectrometry Measurements
Using a Hybrid Significance Testing Approach. Anal Chem 91, 8008–8016 (2019).
}
\seealso{
\code{\link{create_p_diff_uptake_dataset}}
}
