% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlaps.R
\name{plot_overlap}
\alias{plot_overlap}
\title{Plot overlap data}
\usage{
plot_overlap(dat)
}
\arguments{
\item{dat}{data imported by the \code{\link{read_hdx}} function.}
}
\value{
a \code{\link{ggplot2}} object.
}
\description{
Generates overapping peptide plot based 
on supplied data and parameters.
}
\details{
The overlap plot presents all the peptides in given state
on the protein sequence. 
This plot is visible in GUI.
}
\examples{
dat <- read_hdx(system.file(package = "HaDeX", "HaDeX/data/KD_180110_CD160_HVEM.csv"))
plot_overlap(dat)

}
\seealso{
\code{\link{read_hdx}}
\code{\link{show_overlap_data}}
}
