% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlaps.R
\name{plot_overlap}
\alias{plot_overlap}
\title{Plot overlap data}
\usage{
plot_overlap(dat)
}
\arguments{
\item{dat}{produced by \code{\link{show_overlap_data}}
function}
}
\value{
a \code{\link{ggplot2}} object.
}
\description{
Generates overapping peptide plot based 
on supplied data and parameters.
}
\details{
This plot is visible in GUI.
}
\seealso{
\code{\link{read_hdx}}
\code{\link{show_overlap_data}}
}
