% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlaps.R
\name{plot_overlap_distribution}
\alias{plot_overlap_distribution}
\title{Plot overlap distribution}
\usage{
plot_overlap_distribution(dat, start = 1, end = max(dat[["pos"]]))
}
\arguments{
\item{dat}{produced by \code{\link{create_overlap_distribution_dataset}}
function}

\item{start}{start start position of chosen protein.}

\item{end}{end position of chosen protein.}
}
\value{
a \code{\link{ggplot2}} object.
}
\description{
Generates overlap distribution plot based on supplied data
and parameters.
}
\details{
This plot is visible in GUI.
}
\seealso{

}
