% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlaps.R
\name{plot_overlap_distribution}
\alias{plot_overlap_distribution}
\title{generate_overlap_distribution_plot}
\usage{
plot_overlap_distribution(dat, start, end)
}
\arguments{
\item{dat}{produced by \code{\link{generate_overlap_distribution_plot}}
function}

\item{start}{...}

\item{end}{...}
}
\value{
...
}
\description{
Generates overlap distribution plot based on supplied data
and parameters.
}
\details{
This plot is visible in GUI.
}
\seealso{
...
}
