% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_position_frequency.R
\name{plot_position_frequency}
\alias{plot_position_frequency}
\title{Plot position frequency}
\usage{
plot_position_frequency(dat, chosen_state = dat[["State"]][1])
}
\arguments{
\item{dat}{data as imported by the \code{\link{read_hdx}} function}

\item{chosen_state}{sequence states to be included in plot}
}
\description{
Plots the frequency of positions
}
\examples{
dat <- read_hdx(system.file(package = "HaDeX", 
                            "HaDeX/data/KD_180110_CD160_HVEM.csv"))
plot_position_frequency(dat, chosen_state = "CD160_HVEM")

}
