% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_state_comparison.R
\name{plot_state_comparison}
\alias{plot_state_comparison}
\title{State deuterium uptake comparison}
\usage{
plot_state_comparison(
  uptake_dat,
  theoretical = FALSE,
  fractional = FALSE,
  line_size = 1.5,
  all_times = FALSE,
  time_t = NULL
)
}
\arguments{
\item{uptake_dat}{data produced by \code{\link{calculate_state_uptake}} 
function}

\item{theoretical}{\code{logical}, indicator if values are 
calculated using theoretical controls}

\item{fractional}{\code{logical}, indicator if values are shown 
in fractional form}
}
\value{
a \code{ggplot} object
}
\description{
Comparison plot of deuterium uptake values in time
point for biological states.
}
\details{
The function \code{\link{plot_state_comparison}} generates
a comparison plot, presenting deuterium uptake values of peptides 
from selected protein in selected biological states at one time point 
of measurement at once. 
On X-axis there is a position in a sequence, with length of a segment 
of each peptide representing its length. On Y-axis there 
is deuterium uptake in selected form. Error bars represents the combined
and propagated uncertainty.
}
\examples{
dat <- read_hdx(system.file(package = "HaDeX", "HaDeX/data/KD_180110_CD160_HVEM.csv"))
comparison_dat <- calculate_state_uptake(dat)
plot_state_comparison(comparison_dat)

uptake_dat <- create_uptake_dataset(dat, states = c("CD160", "CD160_HVEM"))
plot_state_comparison(uptake_dat, all_times = T)
plot_state_comparison(uptake_dat, fractional = T, all_times = T)

}
\seealso{
\code{\link{read_hdx}}
\code{\link{calculate_state_uptake}}
}
