% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_state_comparison.R
\name{plot_state_comparison}
\alias{plot_state_comparison}
\title{Plot state comparison}
\usage{
plot_state_comparison(
  uptake_dat,
  theoretical = FALSE,
  fractional = FALSE,
  line_size = 1.5,
  all_times = FALSE,
  time_t = NULL
)
}
\arguments{
\item{uptake_dat}{data produced by \code{\link{calculate_state_uptake}} function.}

\item{theoretical}{\code{logical}, determines if values are theoretical.}

\item{fractional}{\code{logical}, determines if values are fractional.}
}
\value{
a \code{ggplot} object.
}
\description{
Plot state comparison
}
\details{
Function \code{\link{plot_state_comparison}} presents provided 
data in a form of comparison plot, for peptides for chosen protein in chosen states,
at one time point of measurement at once. On X-axis there is a position in a sequence, 
with length of a segment of each peptide representing its length. On Y-axis there 
is deuterium uptake in chosen form. Error bars represents the combined and propagated
uncertainty. 
This plot is visible in GUI.
}
\examples{
dat <- read_hdx(system.file(package = "HaDeX", "HaDeX/data/KD_180110_CD160_HVEM.csv"))
comparison_dat <- calculate_state_uptake(dat)
plot_state_comparison(comparison_dat)

uptake_dat <- create_uptake_dataset(dat, states = c("CD160", "CD160_HVEM"))
plot_state_comparison(uptake_dat, all_times = T)

}
\seealso{
\code{\link{calculate_state_uptake}}
}
