% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_uptake_curve.R
\name{plot_uptake_curve}
\alias{plot_uptake_curve}
\title{Deuterium uptake curve}
\usage{
plot_uptake_curve(
  uc_dat,
  theoretical = FALSE,
  fractional = FALSE,
  uncertainty_type = "ribbon",
  log_x = TRUE
)
}
\arguments{
\item{uc_dat}{data produced by \code{\link{calculate_kinetics}} 
or \code{\link{calculate_peptide_kinetics}} or \code{\link{create_kinetic_dataset}}
functions}

\item{theoretical}{\code{logical}, indicator if values are 
calculated using theoretical controls}

\item{fractional}{\code{logical}, indicator if values are shown 
in fractional form}

\item{uncertainty_type}{type of uncertainty presentation, possible values:
"ribbon", "bars" or "bars + line"}

\item{log_x}{\code{logical}, indicator if the X axis values 
are transformed to log10}
}
\value{
a \code{\link{ggplot}} object
}
\description{
Plot deuterium uptake curve for selected peptides
}
\details{
The function \code{\link{plot_uptake_curve}} generates
the deuterium uptake curve for selected peptides 
from selected protein.
On X-axis there are time points of measurements. On Y-axis there
is deuterium uptake in selected form. The combined and propagated
uncertainty can be presentes as ribbons or error bars.
}
\examples{
dat <- read_hdx(system.file(package = "HaDeX", "HaDeX/data/KD_180110_CD160_HVEM.csv"))

uc_dat <- calculate_kinetics(dat, protein = "db_CD160",
                             sequence = "INITSSASQEGTRLN", 
                             state = "CD160",
                             start = 1, end = 15,
                             time_0 = 0.001, time_100 = 1440)
plot_uptake_curve(uc_dat = uc_dat, 
                  theoretical = FALSE, 
                  fractional = TRUE)

uc_dat2 <- calculate_peptide_kinetics(dat, protein = "db_CD160",
                                      sequence = "INITSSASQEGTRLN", 
                                      states = c("CD160", "CD160_HVEM"),
                                      start = 1, end = 15,
                                      time_0 = 0.001, time_100 = 1440)
plot_uptake_curve(uc_dat = uc_dat2, 
                  theoretical = FALSE, 
                  fractional = TRUE)
                
}
\seealso{
\code{\link{read_hdx}}
\code{\link{calculate_kinetics}}
\code{\link{calculate_peptide_kinetics}}
\code{\link{create_kinetic_dataset}}
}
