% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_plots.R
\name{plot_volcano}
\alias{plot_volcano}
\title{Volcano plot}
\usage{
plot_volcano(
  p_dat,
  state_1 = "",
  state_2 = "",
  adjust_axes = TRUE,
  show_confidence_limits = FALSE,
  confidence_level = 0.98,
  color_times = TRUE,
  fractional = F,
  theoretical = F
)
}
\arguments{
\item{p_dat}{data produced by the \code{\link{create_p_diff_uptake_dataset}}
function.}

\item{state_1}{biological state for chosen protein. It is used in the title.}

\item{state_2}{biological state for chosen protein. It is used in the title.}

\item{adjust_axes}{logical, indicating if the X-axis is symmetrical in
relation to 0.}

\item{show_confidence_limits}{logical, indicates if the hybrid testing
confidence intervals are shown.}

\item{confidence_level}{confidence level for the test, from range [0, 1]. It
should be the same as used in \code{\link{create_volcano_dataset}} function.}

\item{color_times}{logical, indicating if different time points on the plot
are distinguishable by color.}

\item{theoretical}{...}

\item{relative}{...}
}
\value{
a \code{\link{ggplot}} object.
}
\description{
Volcano plot
}
\details{
The data produced by \code{\link{create_volcano_dataset}} are plotted
in the form of a volcano plot. The generation of the data is described in the documentation
of \code{\link{create_volcano_dataset}} function. The confidence limit on
P-value is calculated based on the confidence level. The confidence limit on deuterium
uptake difference is calculated using the Houde test for the time point of measurement
from the provided data. The confidence limits are indicated by the red dotted
lines. The points above confidence limits (upper right and left corner) are
statistically significant in hybrid testing.
This plot is visible in GUI.
}
\examples{
dat <-  read_hdx(system.file(package = "HaDeX", "HaDeX/data/KD_180110_CD160_HVEM.csv"))
p_dat <- create_volcano_dataset(dat)
plot_volcano(vol_dat, show_confidence_limits = T)

}
\references{
Hageman, T. S. & Weis, D. D. Reliable Identification of Significant
Differences in Differential Hydrogen Exchange-Mass Spectrometry Measurements
Using a Hybrid Significance Testing Approach. Anal Chem 91, 8008–8016 (2019).

Houde, D., Berkowitz, S.A., and Engen, J.R. (2011).
The Utility of Hydrogen/Deuterium Exchange Mass Spectrometry in
Biopharmaceutical Comparability Studies. J Pharm Sci 100, 2071–2086.
}
\seealso{
\code{\link{create_p_diff_uptake_dataset}}
\code{\link{show_volcano_data}}
}
