% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_control.R
\name{quality_control}
\alias{quality_control}
\title{Experiment quality control}
\usage{
quality_control(
  dat,
  protein,
  state_first,
  state_second,
  time_t,
  time_0,
  deut_part
)
}
\arguments{
\item{dat}{data read by \code{\link{read_hdx}}}

\item{protein}{...}

\item{state_first}{state of the first peptide}

\item{state_second}{state of the second peptide}

\item{time_t}{chosen time point}

\item{time_0}{`in` time}

\item{deut_part}{...}
}
\value{
\code{data.frame} with mean uncertainty per different `out_time` value. The values are shown as percentages.
}
\description{
Checks how the uncertainty changes in a function of `out_time`.
}
\details{
The function calculates mean uncertainty of all peptides and its uncertainty (standard error) based on given `in_time` and `chosen_time` 
as a function of `out_time`. Both theoretical and experimental results for each state and their difference are supplied for comparison but only 
experimental calculations depends on `out_time` variable. The results are either in form of relative or absolute values depending on the `relative` 
parameter supplied by the user. 
This data can be useful for general overview of the experiment and analyse of the chosen time parameters.
}
\examples{
# load example data
dat <- read_hdx(system.file(package = "HaDeX", "HaDeX/data/KD_180110_CD160_HVEM.csv"))

# calculate mean uncertainty 
(result <- quality_control(dat = dat,
                           protein = "db_CD160",
                           state_first = "CD160",
                           state_second = "CD160_HVEM", 
                           time_t = 1, 
                           time_0 = 0.001,
                           deut_part = 1))    
                           
# load extra libraries
library(ggplot2)
library(tidyr)
library(dplyr)

# example of data visualization 
gather(result, 2:7, key = 'type', value = 'value') \%>\%
filter(startsWith(type, "avg")) \%>\%
  ggplot(aes(x = factor(out_time), y = value, group = type)) +
  geom_line(aes(color = type)) +
  labs(x = "Out time", 
       y = "Mean uncertainty")

}
\seealso{
\code{\link{read_hdx}}
}
