% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_hdx.R
\name{read_hdx}
\alias{read_hdx}
\title{Read HDX-MS data file}
\usage{
read_hdx(filename)
}
\arguments{
\item{filename}{a file supplied by a user. Formats allowed: .csv, .tsv and .xls.}
}
\value{
a \code{data.frame}.
}
\description{
Imports data from a HDX-MS file (as provided by Waters DynamX software) 
and validates its content.
}
\examples{
dat <- read_hdx(system.file(package = "HaDeX", 
                            "HaDeX/data/KD_180110_CD160_HVEM.csv"))

}
