% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_hdx.R
\name{read_hdx}
\alias{read_hdx}
\title{read_hdx}
\usage{
read_hdx(filename)
}
\arguments{
\item{filename}{file supplied by user}
}
\value{
data frame for further modifications
}
\description{
Imports data from file and validates its content.
}
