% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconstruct_sequence.R
\name{reconstruct_sequence}
\alias{reconstruct_sequence}
\title{reconstruct_sequence}
\usage{
reconstruct_sequence(dat)
}
\arguments{
\item{dat}{data frame with data from Dynamix file}
}
\value{
reconstructed sequence (character)
}
\description{
Reconstructs protein sequence from DynamiX file. If a position is not covered by experiment data, x is shown.
}
\examples{
dat <- read_hdx(system.file(package = "HaDeX", "HaDeX/data/KD_180110_CD160_HVEM.csv"))
reconstruct_sequence(dat)

}
