% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconstruct_sequence.R
\name{reconstruct_sequence}
\alias{reconstruct_sequence}
\title{Reconstruct protein sequence}
\usage{
reconstruct_sequence(
  dat,
  protein = dat[["Protein"]][1],
  states = unique(dat[["State"]]),
  end = NULL
)
}
\arguments{
\item{dat}{data read by \code{\link{read_hdx}}}

\item{protein}{the protein of which the sequence is reconstructed}

\item{states}{biological states, for which the sequence is reconstructed
and coverage calculated}

\item{end}{\code{\link{numeric}}, end position of the protein, optional.
If not provided, is read from the file. It allows to prelongate the sequence
if the end of the sequence is cut.}
}
\value{
reconstructed sequence - \code{character} object.
}
\description{
Reconstructs protein sequence from supplied file.
}
\details{
The function reconstructs protein sequence from supplied 
experimental data. If a position is not covered, x is shown.
First version doesn't support manual sequence length correction.
}
\examples{
dat <- read_hdx(system.file(package = "HaDeX", "HaDeX/data/KD_180110_CD160_HVEM.csv"))
reconstruct_sequence(dat)

}
\seealso{
\code{\link{read_hdx}}
}
