% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_data.R
\name{show_diff_uptake_data}
\alias{show_diff_uptake_data}
\title{Show differential uptake data}
\usage{
show_diff_uptake_data(
  diff_uptake_dat,
  theoretical = FALSE,
  fractional = FALSE,
  renamed = TRUE
)
}
\arguments{
\item{diff_uptake_dat}{data produced by 
\code{\link{create_diff_uptake_dataset}} function.}

\item{theoretical}{\code{logical}, determines if values are theoretical.}

\item{fractional}{\code{logical}, determines if values are fractional.}

\item{renamed}{\code{logical}, determines if the names of the columns
are renamed to user-friendly ones. Currently FALSE not implemented.}
}
\value{
a \code{\link{data.frame}} object.
}
\description{
Show differential uptake data
}
\details{
This function subsets the dataset based on provided criteria,
rounds the numerical values (4 places) and changes the column names 
to user-friendly ones. 
This data is available in the GUI.
}
\examples{
dat <- read_hdx(system.file(package = "HaDeX", "HaDeX/data/KD_180110_CD160_HVEM.csv"))
diff_uptake_dat <- create_diff_uptake_dataset(dat)
head(show_diff_uptake_data(diff_uptake_dat))

}
\seealso{
\code{\link{read_hdx}}
\code{\link{create_diff_uptake_dataset}}
}
