% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_data.R
\name{show_diff_uptake_data_confidence}
\alias{show_diff_uptake_data_confidence}
\title{Show differential uptake data with confidence levels}
\usage{
show_diff_uptake_data_confidence(
  dat,
  theoretical = FALSE,
  fractional = FALSE,
  confidence_level_1 = 0.98,
  confidence_level_2 = 0.99
)
}
\arguments{
\item{dat}{data produced by \code{\link{generate_differential_data_set}}
function.}

\item{theoretical}{\code{logical}, determines if values are theoretical.}

\item{fractional}{\code{logical}, determines if values are fractional.}

\item{confidence_limit_1}{...}

\item{confidence_limit_2}{...}
}
\value{
a \code{\link{data.frame}} object.
}
\description{
Show differential uptake data with confidence levels
}
\details{
This function subsets the dataset based on provided criteria,
rounds the numerical values (4 places) and changes the column names 
to user-friendly ones. 
This data is available in the GUI.
}
\examples{
dat <- read_hdx(system.file(package = "HaDeX", "HaDeX/data/KD_180110_CD160_HVEM.csv"))
diff_dat <- generate_differential_data_set(dat)
show_diff_uptake_data_confidence(diff_dat)

}
\seealso{
\code{\link{generate_differential_data_set}}
\code{\link{plot_differential}}
}
