% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_kinetic_elements.R
\name{show_kinetic_data}
\alias{show_kinetic_data}
\title{generate_kinetic_data}
\usage{
show_kinetic_data(dat, theoretical = FALSE, fractional = FALSE)
}
\arguments{
\item{dat}{custom format, produced by 
\code{\link{generate_kinetic_data_set}}}

\item{theoretical}{\code{logical}, determines if plot shows theoretical values}

\item{fractional}{\code{logical}, determines if plot shows fractional values}
}
\value{
...
}
\description{
Generates deuterium uptake data, based on the supplied
parameters.
}
\details{
This data is available in the GUI. 
All of the numerical values are rounded to 4 places after the dot!!
}
\seealso{
...
}
