% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlaps.R
\name{show_overlap_data}
\alias{show_overlap_data}
\title{show_overlap_data}
\usage{
show_overlap_data(dat, protein, state, start, end)
}
\arguments{
\item{dat}{...}

\item{protein}{...}

\item{state}{...}

\item{start}{...}

\item{end}{...}
}
\value{
...
}
\description{
Generates overlap data, based on the supplied
parameters.
}
\details{
This data is available in the GUI.
}
\seealso{
...
}
