% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlaps.R
\name{show_overlap_data}
\alias{show_overlap_data}
\title{Show data on overlap}
\usage{
show_overlap_data(dat, protein, state, start, end)
}
\arguments{
\item{dat}{data imported by the \code{\link{read_hdx}} function.}

\item{protein}{chosen protein.}

\item{state}{biological state for chosen protein.}

\item{start}{start start position of chosen protein.}

\item{end}{end position of chosen protein.}
}
\value{
a \code{\link{data.frame}} object.
}
\description{
Presents overlap data, based on the supplied
parameters.
}
\details{
This data is available in the GUI.
}
\seealso{
\code{\lnk{read_hdx}}
}
