% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_control.R
\name{show_quality_control_data}
\alias{show_quality_control_data}
\title{generate_quality_control_data}
\usage{
show_quality_control_data(dat)
}
\arguments{
\item{dat}{data produced by \code{\link{create_quality_control_dataset}} 
function, scaled if necessary.}
}
\value{
a \code{\link{data.frame}} object.
}
\description{
Generates quality control data, based on the supplied
parameters.
}
\details{
This data is available in the GUI. The names of the parameters
and variables will be changed later after the glossary project.
}
\seealso{
\code{\link{create_quality_control_dataset}}
\code{\link{plot_quality_control}}
}
