% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_data.R
\name{show_volcano_data}
\alias{show_volcano_data}
\title{Show volcano data}
\usage{
show_volcano_data(vol_data)
}
\arguments{
\item{vol_data}{data produced by the \code{\link{create_volcano_dataset}} 
function.}
}
\value{
a \code{\link{data.frame}} object.
}
\description{
Show volcano data
}
\details{
This function subsets the dataset based on provided criteria,
rounds the numerical values (4 places) and changes the column names 
to user-friendly ones. 
This data is available in the GUI.
}
\examples{
dat <-  dat <- read_hdx(system.file(package = "HaDeX", "HaDeX/data/KD_180110_CD160_HVEM.csv"))
vol_dat <- create_volcano_dataset(dat)
head(show_volcano_data(vol_dat))

}
\seealso{
\code{\link{create_volcano_dataset}} 
\code{\link{plot_volcano}}
}
