% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_hdexaminer_file.R
\name{update_hdexaminer_file}
\alias{update_hdexaminer_file}
\title{Update data from HDeXaminer}
\usage{
update_hdexaminer_file(
  dat,
  fd_time,
  old_protein_name = NULL,
  new_protein_name = NULL,
  old_state_name = NULL,
  new_state_name = NULL,
  confidence = c("High", "Medium")
)
}
\arguments{
\item{dat}{a data frame with data read from file by \code{\link{read_hdx}}}

\item{fd_time}{time point (in minutes) for fully deuterated sample. Numeric value, bigger than other
time points in the file. Obligatory}

\item{old_protein_name}{protein name to be changed. String or string vector (if the file contains
data from multiple proteins). Not obligatory}

\item{new_protein_name}{new names for old_protein_name. String or string vector (if the file contains
data from multiple proteins). Not obligatory}

\item{old_state_name}{state names to be changed. String or string vector (if the file contains
data from multiple states). Not obligatory}

\item{new_state_name}{new names for old_state_name. String or string vector (if the file contains
data from multiple states). Not obligatory}

\item{confidence}{vector of accepted confidence values (internal flag from HDeXaminer). By default only
accepted values are `Medium` and `High`, with `Low` excluded. The user can control that by specifying 
this parameter}
}
\value{
\code{dat} - a \code{\link{data.frame}} with validated and updated content.
}
\description{
Updates data read from file with detected data source HDeXaminer
}
\details{
Data from HDeXaminer is condensed and automated data retrieving may be corrected by the user. 
The original file has a mark "FD" for fully deuterated data instead of numerical value for time point 
(provided in minutes) that is not consistent for workflow and not enough for precise data description.
Moreover, the data about both protein and state is included in one column and for detailed information
function \code{\link{update_hdexaminer_file}} allows to change them.
}
\examples{
# read example data 
dat_hdexaminer <- read_hdx(system.file(package = "HaDeX", 
                     "HaDeX/data/KD_180110_CD160_HVEM.csv"))
                     

}
\seealso{
\code{\link{read_hdx}} 
\code{\link{calculate_kinetics}} 
\code{\link{plot_coverage}} 
\code{\link{plot_position_frequency}}
\code{\link{reconstruct_sequence}}
}
