% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/woods_plot.R
\name{woods_plot}
\alias{woods_plot}
\title{woods_plot}
\usage{
woods_plot(calc_dat, theoretical = FALSE, relative = TRUE)
}
\arguments{
\item{calc_dat}{processed data from DynamiX file - using prepare_dataset}

\item{theoretical}{logical value to determine if plot is theoretical or not. default : false}

\item{relative}{logical value to determine if values are relative or absolute. default : true}
}
\description{
Produces Woods' plot based on previously processed data - theoretical or experimental. User can change labels if needed.
}
\examples{
dat <- read_hdx(system.file(package = "HaDeX", "HaDeX/data/KD_180110_CD160_HVEM.csv"))
calc_dat <- prepare_dataset(dat,
                            in_state_first = "CD160_0.001",
                            chosen_state_first = "CD160_1",
                            out_state_first = "CD160_1440",
                            in_state_second = "CD160_HVEM_0.001",
                            chosen_state_second = "CD160_HVEM_1",
                            out_state_second = "CD160_HVEM_1440") 
woods_plot(calc_dat = calc_dat,
           theoretical = TRUE)
woods_plot(calc_dat = calc_dat)

}
